% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{add_channel_types}
\alias{add_channel_types}
\title{Add channel types info}
\usage{
add_channel_types(
  data,
  auto = TRUE,
  eog = NULL,
  ecg = NULL,
  photo_sens = NULL,
  df_list = NULL
)
}
\arguments{
\item{data}{an \code{eegUtils} type of EEG object.}

\item{auto}{If none of the other parameters are mentioned except the data,
attempts to determine the channel type from their name. Defaults to \code{TRUE}.
Other parameters are ignored if this is true. Currently only identifies,
EOG and ECG channels. Automatically turns \code{FALSE} in case other arguments
are provided}

\item{eog}{Character vector mentioning the name of EOG channels.}

\item{ecg}{Character vector mentioning the name of ECG channels.}

\item{photo_sens}{Character vector mentioning the name of Photo Sensor channel.}

\item{df_list}{A 2 column data frame mentioning the electrode name
in the first column and the type in second column OR a named list with
names as channel types and values as channel names}
}
\value{
The same type of \code{eegUtils} object with the channel types updated
}
\description{
Adds info about types of channel - EEG/EOG/ECG/Photo Sensor, etc.
}
