% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{add_head}
\alias{add_head}
\title{Add a cartoon head to a topographical plot}
\usage{
add_head(r = 1, colour = "black", scaling = 1)
}
\arguments{
\item{r}{Radius of head}

\item{colour}{Colour of lines}

\item{scaling}{Scaling factor for line width}
}
\description{
Add a cartoon head to a topographical plot
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
