% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{apply_ica}
\alias{apply_ica}
\alias{apply_ica.eeg_ICA}
\alias{apply_ica.eeg_epochs}
\title{Recreate channel timecourses from ICA decompositions.}
\usage{
apply_ica(data, ...)

\method{apply_ica}{eeg_ICA}(data, comps = NULL, ...)

\method{apply_ica}{eeg_epochs}(data, decomp, comps, ...)
}
\arguments{
\item{data}{An \code{eeg_ICA} or \code{eeg_epochs} object.}

\item{...}{Other parameters.}

\item{comps}{Components to remove.}

\item{decomp}{An \code{eeg_ICA} object.}
}
\description{
This function can be used to either recreate "mixed" (i.e. channel level)
timecourses from an ICA decomposition, or to apply a set of ICA weights to a
given dataset for the purpose of removing specific ICA components from that
dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{apply_ica(eeg_ICA)}: From given \code{eeg_ICA} object, recreate channel timecourses.

\item \code{apply_ica(eeg_epochs)}: Combine a specific set of ICA weights with any
\code{eeg_epochs} object.

}}
\examples{
test_ica <- run_ICA(demo_epochs, pca = 10)
plot_butterfly(demo_epochs)
# Reconstruct the original data from the ICA decomposition.
# Note that the ICA process subtracts the mean from each epoch,
# so the reconstructed plot may look slightly different to the original.
plot_butterfly(apply_ica(test_ica))
# Remove component 2 from the data
plot_butterfly(apply_ica(demo_epochs, test_ica, comps = 2))
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
