% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_for_ica.R
\name{ar_acf}
\alias{ar_acf}
\alias{ar_acf.eeg_ICA}
\title{Detect low autocorrelation of ICA components}
\usage{
ar_acf(data, ...)

\method{ar_acf}{eeg_ICA}(data, ms = 20, plot = TRUE, verbose = TRUE, threshold = NULL, ...)
}
\arguments{
\item{data}{\code{eeg_ICA} object}

\item{...}{additional parameters}

\item{ms}{Time lag to check ACF, in milliseconds. Defaults to 20 ms.}

\item{plot}{Produce plot showing ACF and threshold for all EEG components.}

\item{verbose}{Print informative messages. Defaults to TRUE.}

\item{threshold}{Specify a threshold for low ACF. NULL estimates the threshold automatically.}
}
\value{
A character vector of component names that break the threshold.
}
\description{
Low autocorrelation can be a sign of a poor quality channel or component.
Often these are noisy, poor contact, or heavily contaminated with muscle
noise. Low autocorrelation at a lag of 20ms is often associated with muscle
noise.
}
\section{Methods (by class)}{
\itemize{
\item \code{ar_acf(eeg_ICA)}: Autocorrelation checker for \code{eeg_ICA} objects

}}
\examples{
demo_sobi <- run_ICA(demo_epochs, pca = 10)
ar_acf(demo_sobi)
}
\references{
Chaumon, M., Bishop, D.V., Busch, N.A. (2015). A practical guide
to the selection of independent components of the electroencephalogram for
artifact correction. J Neurosci Methods. Jul 30;250:47-63. doi:
10.1016/j.jneumeth.2015.02.025
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
