% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{ar_chanfoc}
\alias{ar_chanfoc}
\title{Detect high channel focality of ICA components}
\usage{
ar_chanfoc(
  data,
  plot = TRUE,
  threshold = NULL,
  verbose = TRUE,
  measure = "max",
  ...
)
}
\arguments{
\item{data}{\code{eeg_ICA} object}

\item{plot}{Produce plot showing max z-scores and threshold for all ICA
components.}

\item{threshold}{Specify a threshold for high focality. NULL estimates the
threshold automatically.}

\item{verbose}{Print informative messages.}

\item{measure}{Use maximum "max" or "kurtosis".}

\item{...}{additional parameters}
}
\description{
Detect components that load heavily on a single channel. Looks for components
that have one particular channel that has a particularly high z-score.
}
\references{
Chaumon, M., Bishop, D.V., Busch, N.A. (2015). A practical guide
  to the selection of independent components of the electroencephalogram for
  artifact correction. J Neurosci Methods. Jul 30;250:47-63. doi:
  10.1016/j.jneumeth.2015.02.025
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
