% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R, R/deprecated.R
\name{ar_FASTER}
\alias{ar_FASTER}
\alias{ar_FASTER.eeg_epochs}
\alias{eeg_FASTER}
\title{FASTER EEG artefact rejection}
\usage{
ar_FASTER(data, ...)

\method{ar_FASTER}{eeg_epochs}(
  data,
  exclude = NULL,
  test_chans = TRUE,
  test_epochs = TRUE,
  test_cine = TRUE,
  ...
)

eeg_FASTER(data, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{...}{Parameters passed to FASTER}

\item{exclude}{Channels to be ignored by FASTER.}

\item{test_chans}{Logical. Run tests of global channel statistics}

\item{test_epochs}{Logical. Run tests of globally bad epochs.}

\item{test_cine}{Logical. Run tests for locally bad channels within epochs.}
}
\value{
An \code{eeg_epochs} object with artefact correction applied.
}
\description{
An implementation of the FASTER artefact rejection method for EEG by Nolan,
Whelan & Reilly (2010) FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods. Not yet fully implemented.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Run FASTER on \code{eeg_epochs}
}}

\examples{
ar_FASTER(demo_epochs)
}
\references{
Nolan, Whelan & Reilly (2010). FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
