% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{ar_trialfoc}
\alias{ar_trialfoc}
\title{Detect high trial focality of ICA components}
\usage{
ar_trialfoc(data, plot = TRUE, threshold = NULL, verbose = TRUE)
}
\arguments{
\item{data}{\code{eeg_ICA} object}

\item{plot}{Produce plot showing max z-scores and threshold for all ICA
components.}

\item{threshold}{Specify a threshold (z-score) for high focality. NULL estimates the
threshold automatically.}

\item{verbose}{Print informative messages.}
}
\description{
Detect components that load heavily on a small number of trials. Looks for components
that have one particular trial that has a particularly high z-score.
}
\references{
Chaumon, M., Bishop, D.V., Busch, N.A. (2015). A practical guide
  to the selection of independent components of the electroencephalogram for
  artifact correction. J Neurosci Methods. Jul 30;250:47-63. doi:
  10.1016/j.jneumeth.2015.02.025
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
