% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{array_t}
\alias{array_t}
\title{Array one-sample t-statistic}
\usage{
array_t(x, mu = 0)
}
\arguments{
\item{x}{3D array (timepoints x epochs x electrodes)}

\item{mu}{mean to test against}
}
\description{
Array one-sample t-statistic
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
