% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_data}
\alias{as.data.frame.eeg_data}
\title{Generic as.data.frame method for objects of class \code{eeg_data}}
\usage{
\method{as.data.frame}{eeg_data}(x, long = FALSE)
}
\description{
Generic as.data.frame method for objects of class \code{eeg_data}
}
