% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_epochs}
\alias{as.data.frame.eeg_epochs}
\title{Convert \code{eeg_epochs} object to data.frame}
\usage{
\method{as.data.frame}{eeg_epochs}(x, row.names = NULL,
  optional = FALSE, long = FALSE, events = FALSE,
  conditions = NULL, cond_labels = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{eeg_epochs}}

\item{row.names}{Kept for compatability with S3 generic, ignored.}

\item{optional}{Kept for compatability with S3 generic, ignored.}

\item{long}{Convert to long format. Defaults to FALSE.}

\item{events}{Include events in output. Defaults to FALSE.}

\item{conditions}{Column names of}

\item{cond_labels}{Add column tagging epochs with events that have matching
labels.}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert an \code{eeg_epochs} object to a data.frame for use with whatever
packages you desire.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
