% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_ICA}
\alias{as.data.frame.eeg_ICA}
\title{Convert \code{eeg_ICA} object to data frame}
\usage{
\method{as.data.frame}{eeg_ICA}(x, row.names = NULL, optional = FALSE,
  long = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{eeg_ICA}}

\item{row.names}{Kept for compatability with S3 generic, ignored.}

\item{optional}{Kept for compatability with S3 generic, ignored.}

\item{long}{Convert to long format. Defaults to FALSE}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert \code{eeg_ICA} object to data frame
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
