% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_converters.R
\name{as.data.frame.eeg_lm}
\alias{as.data.frame.eeg_lm}
\title{Convert \code{eeg_lm} to data.frame}
\usage{
\method{as.data.frame}{eeg_lm}(
  x,
  row.names = NULL,
  optional = FALSE,
  long = FALSE,
  coords = TRUE,
  values = c("coefficients", "std_err", "t_stats", "r_sq"),
  ...
)
}
\arguments{
\item{x}{Object of class \code{eeg_data}}

\item{row.names}{Kept for compatibility with S3 generic, ignored.}

\item{optional}{Kept for compatibility with S3 generic, ignored.}

\item{long}{Convert to long format. Defaults to FALSE.}

\item{coords}{Include electrode coordinates in output. Only possible when
long = TRUE.}

\item{values}{Defaults to "coefficients", returning fitted coefficient values.}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert an object of class \code{eeg_data} into a standard \code{data.frame}.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
