% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_stats}
\alias{as.data.frame.eeg_stats}
\title{Convert \code{eeg_stats} objects to data frames}
\usage{
\method{as.data.frame}{eeg_stats}(
  x,
  row.names = NULL,
  optional = FALSE,
  long = FALSE,
  coords = FALSE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{eeg_stats}}

\item{row.names}{Kept for compatability with S3 generic, ignored.}

\item{optional}{Kept for compatability with S3 generic, ignored.}

\item{long}{Convert to long format. Defaults to FALSE.}

\item{coords}{Include electrode coordinates in output (ignored if long = FALSE)}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert \code{eeg_stats} objects to data frames
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
