% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{average_tf}
\alias{average_tf}
\title{Internal function for averaging over epochs}
\usage{
average_tf(data)
}
\arguments{
\item{data}{data to average over}
}
\description{
Internal function for averaging over epochs
}
\keyword{internal}
