% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_ica_ar.R
\name{bad_epochs}
\alias{bad_epochs}
\title{Identify globally bad epochs}
\usage{
bad_epochs(data, sds = 2, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{sds}{Standard deviation thresholds}

\item{...}{Further parameters (tbd)}
}
\value{
A numeric vector with epoch numbers
}
\description{
Identify bad epochs that should be excluded from ICA decomposition.
Uses standard deviation to identify bad epochs
\emph{See Reference for more details}
}
\references{
Nolan, Whelan & Reilly (2010). FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods.
}
\seealso{
\code{\link[=bad_chans]{bad_chans()}}

Other Pre-ICA artifact rejection: 
\code{\link{bad_chans}()}
}
\author{
Akash Mer
}
\concept{Pre-ICA artifact rejection}
