% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{biharm_topo}
\alias{biharm_topo}
\title{Create a biharmonic smooth across the scalp}
\usage{
biharm_topo(data, scaled_x, scaled_y, grid_res)
}
\arguments{
\item{data}{Data frame from which to generate predictions}

\item{scaled_x}{x coordinates rescaled}

\item{scaled_y}{y coordinates rescaled}

\item{grid_res}{resolution of output grid}
}
\description{
Create a biharmonic smooth across the scalp
}
\keyword{internal}
