% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{bva_elecs}
\alias{bva_elecs}
\title{Convert BVA spherical locations}
\usage{
bva_elecs(chan_info, circumference = 85)
}
\arguments{
\item{chan_info}{A data.frame containing electrodes}

\item{circumference}{Head circumference in millimetres}
}
\description{
Reads the BrainVoyager spherical electrode locations and converts them to
Cartesian 3D and 2D projections.
}
\keyword{internal}
