% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{calc_fstat}
\alias{calc_fstat}
\title{Calculate F-statistic for independent samples}
\usage{
calc_fstat(x1, x2)
}
\arguments{
\item{x1}{condition one}

\item{x2}{condition two}
}
\description{
Calculate F-statistic for independent samples
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
\keyword{internal}
