% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_max_elec}
\alias{calc_max_elec}
\title{Calculate maximum electrode distance from origin.}
\usage{
calc_max_elec(data)
}
\description{
Calculate maximum electrode distance from origin.
}
\keyword{internal}
