% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{calc_pval}
\alias{calc_pval}
\title{Calculate p-value}
\usage{
calc_pval(x, df, tails = 2)
}
\arguments{
\item{x}{t-stat}

\item{df}{Degrees of freedom}

\item{tails}{1 or 2 for one- or two-tailed p-value.}
}
\description{
Calculate p-value
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
