% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{calc_tstat}
\alias{calc_tstat}
\title{Calculate t-statistic}
\usage{
calc_tstat(x, mu = 0)
}
\arguments{
\item{x}{vector of values to compare against mu}

\item{mu}{mean to compare x to.}
}
\description{
Calculate t-statistic
}
