% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{calc_tstat_2}
\alias{calc_tstat_2}
\title{Calculate two-sample independent t-statistic using Welch's formula}
\usage{
calc_tstat_2(x1, x2)
}
\arguments{
\item{x1}{vector of values from one condition}

\item{x2}{vector of values from one condition}
}
\description{
Calculate two-sample independent t-statistic using Welch's formula
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
