% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_eeg.R
\name{calculate_sem}
\alias{calculate_sem}
\title{Calculate the standard error of the mean}
\usage{
calculate_sem(data, time_lim = NULL, conditions = NULL)
}
\arguments{
\item{data}{An \code{eegUtils} object}

\item{time_lim}{A vector of time-points to calculate SEM over.}

\item{conditions}{Split the calculation across multiple conditions}
}
\description{
Calculate the standard error of the mean
}
\keyword{internal}
