% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{cart_to_sph}
\alias{cart_to_sph}
\title{Convert 3D Cartesian co-ordinates to spherical}
\usage{
cart_to_sph(x, y, z)
}
\arguments{
\item{x}{X co-ordinates}

\item{y}{Y co-ordinates}

\item{z}{Z co-ordinates}
}
\value{
Data frame with entries "sph_radius",
 "sph_phi" (in degrees),
  "sph_theta" (in degrees).
}
\description{
Convert 3D Cartesian co-ordinates to spherical
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
