% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{cart_to_spherical}
\alias{cart_to_spherical}
\title{Convert 3D Cartesian coordinates to spherical coordinates}
\usage{
cart_to_spherical(xyz_coords)
}
\arguments{
\item{xyz_coords}{3D Cartesian electrode locations}
}
\description{
Output theta and phi are in degrees.
}
\keyword{internal}
