% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{channel_names}
\alias{channel_names}
\title{Retrieve signal/channel names}
\usage{
channel_names(.data)
}
\arguments{
\item{.data}{\code{eegUtils object}}
}
\description{
Get the names of the `signals` element of `eegUtils` objects.
}
\examples{
channel_names(demo_epochs)

}
