% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{channel_stats}
\alias{channel_stats}
\title{Channel statistics}
\usage{
channel_stats(data, ...)
}
\arguments{
\item{data}{Data as a \code{eeg_data} or \code{eeg_epochs} object.}

\item{...}{Other parameters passed to the functions.}
}
\description{
Channel statistics
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
