% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{channels}
\alias{channels}
\alias{channels<-}
\title{Modify channel information}
\usage{
channels(.data)

channels(.data) <- value
}
\arguments{
\item{.data}{\code{eegUtils} object to view}

\item{value}{Value to replace `chan_info` structure with.}
}
\description{
Get or set the contents of the channel information inside \code{eegUtils} objects.
}
\examples{
channels(demo_epochs)

}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
