% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{check_ci_str}
\alias{check_ci_str}
\title{Check if chan_info is in old format}
\usage{
check_ci_str(chan_info)
}
\arguments{
\item{chan_info}{Channel info structure}
}
\description{
Check if chan_info is in old format
}
\keyword{internal}
