% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_averaging.R
\name{check_classes}
\alias{check_classes}
\title{Check that all classes in a list match}
\usage{
check_classes(data)
}
\arguments{
\item{data}{list of objects to check}
}
\description{
Check that all classes in a list match
}
\keyword{internal}
