% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_items.R
\name{check_epochs}
\alias{check_epochs}
\title{Check that all objects have the same number of trials}
\usage{
check_epochs(data)
}
\arguments{
\item{data}{list of objects to check}
}
\description{
Check that all objects have the same number of trials
}
\keyword{internal}
