% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{check_q}
\alias{check_q}
\title{Validate the q factor for downsampling}
\usage{
check_q(q, srate)
}
\arguments{
\item{q}{Q factor}

\item{srate}{Sampling rate}
}
\description{
Validate the q factor for downsampling
}
\keyword{internal}
