% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{check_timings}
\alias{check_timings}
\alias{check_timings.eeg_data}
\alias{check_timings.eeg_epochs}
\title{Check consistency of event and timing tables}
\usage{
check_timings(.data)

\method{check_timings}{eeg_data}(.data)

\method{check_timings}{eeg_epochs}(data)
}
\arguments{
\item{data}{\code{eeg_data} or \code{eeg_epochs} object}
}
\description{
Check consistency of event and timing tables
}
\keyword{internal}
