% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{circ_mean}
\alias{circ_mean}
\title{Calculate circular mean}
\usage{
circ_mean(data)
}
\arguments{
\item{data}{vector of phase angles in radians}
}
\description{
Calculates the circular mean from vector of phase angles. Input should be in
radians.
}
\keyword{internal}
