% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{compare_epochs}
\alias{compare_epochs}
\alias{compare_epochs.eeg_epochs}
\title{Compare epochs using an independent t-test}
\usage{
compare_epochs(data, ...)

\method{compare_epochs}{eeg_epochs}(data, cond_label = NULL, type, ...)
}
\arguments{
\item{data}{\code{eeg_epochs} object.}

\item{...}{Other parameters passed to functions.}

\item{cond_label}{Conditions to test. A character vector of max length 2.}

\item{type}{Type of test to use. "1samp", "2samp"}
}
\description{
Compare epochs using an independent t-test
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Compare differences across epochs
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
