% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{compute_csd}
\alias{compute_csd}
\title{Calculate current source densities}
\usage{
compute_csd(data, m = 4, smoothing = 1e-05, scaling = 1)
}
\arguments{
\item{data}{\code{eeg_data} object}

\item{m}{smoothing constraint (higher = more rigid)}

\item{smoothing}{lambda constant}

\item{scaling}{Default scaling (1) is uV / m^2.}
}
\description{
Calculate current source densities
}
\keyword{internal}
