% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{compute_g}
\alias{compute_g}
\title{Compute the g function for two sets of locations of channel locations on the unit sphere.}
\usage{
compute_g(xyz_coords, xyz_elecs)
}
\arguments{
\item{xyz_coords}{A set of electrodes locations on a unit sphere.}

\item{xyz_elecs}{A set of electrodes locations on a unit sphere.}
}
\description{
Compute the g function for two sets of locations of channel locations on the unit sphere.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
