% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{compute_h}
\alias{compute_h}
\title{Compute the h function for two sets of locations of channel locations on the
unit sphere.}
\usage{
compute_h(xyz_coords, xyz_elecs, m = 4, iter = 50)
}
\arguments{
\item{xyz_coords}{A set of electrode locations on a unit sphere.}

\item{xyz_elecs}{A set of electrode locations on a unit sphere.}

\item{m}{Interpolation constant (higher = less flexible)}

\item{iter}{iterations for calculations}

\item{lambda}{smoothing parameter}
}
\description{
Compute the h function for two sets of locations of channel locations on the
unit sphere.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
