% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{compute_tfr}
\alias{compute_tfr}
\alias{compute_tfr.default}
\alias{compute_tfr.eeg_epochs}
\title{Compute Time-Frequency representation of EEG data}
\usage{
compute_tfr(data, ...)

\method{compute_tfr}{default}(data, ...)

\method{compute_tfr}{eeg_epochs}(data, method = "morlet", foi, n_freq,
  n_cycles = 7, keep_trials = TRUE, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}.}

\item{...}{Furthere TFR parameters}

\item{method}{Time-frequency analysis method. Defaults to "morlet".}

\item{foi}{Frequencies of interest. Scalar or character vector of the lowest
and highest frequency to resolve.}

\item{n_freq}{Number of frequencies to be resolved.}

\item{n_cycles}{Number of cycles at each frequency.}

\item{keep_trials}{Keep single trials or average over them before returning.}
}
\description{
This function creates a time frequency represention of EEG time series data.
Currently, the only available method is a Morlet wavelet transformation
performed using convolution in the frequency domain.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for compute_tfr

\item \code{eeg_epochs}: Default method for compute_tfr
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
