% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{cov_epochs}
\alias{cov_epochs}
\title{Covariance of epoched data}
\usage{
cov_epochs(data)
}
\arguments{
\item{data}{epoched data to calculate covariance for}
}
\description{
Calculate covariance of each epoch, then average
}
\keyword{internal}
