% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{create_electrode_layout_plot}
\alias{create_electrode_layout_plot}
\title{Create a ggplot or plotly plot showing electrode positions}
\usage{
create_electrode_layout_plot(data, interact)
}
\arguments{
\item{data}{The data from which to extract channel locations}

\item{interact}{If TRUE, use \code{plotly} to create a 3D interactive plot, otherwise create a 2D \code{ggplot}.}
}
\description{
Create a ggplot or plotly plot showing electrode positions
}
\keyword{internal}
