% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_images.R
\name{create_erpimage}
\alias{create_erpimage}
\title{Function for creating an ERP image}
\usage{
create_erpimage(data, electrode, smoothing, clim, interpolate = FALSE)
}
\arguments{
\item{data}{Data frame to be plotted. Requires an amplitude column.}

\item{electrode}{electrode at which to generate an ERP image.}

\item{smoothing}{Number of trials to smooth over when generating image}

\item{clim}{Character vector of min and max values of plotting colour range.
e.g. c(-5,5). Defaults to min and max.}

\item{interpolate}{Turn on geom_raster() interpolation for smoother images.}
}
\description{
Function for creating an ERP image
}
\keyword{internal}
