% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_averaging.R
\name{create_grandavg}
\alias{create_grandavg}
\title{Create a grand average file from multiple evoked objects}
\usage{
create_grandavg(data, keep_indivs, x = NULL)
}
\description{
Create a grand average file from multiple evoked objects
}
\keyword{internal}
