% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{create_head}
\alias{create_head}
\title{Create head shape for plotting}
\usage{
create_head(r, circ_rads)
}
\arguments{
\item{r}{Radius of head}

\item{circ_rads}{Circle outline in radians}
}
\description{
Creates data frames for plotting a head with ears and nose.
}
\keyword{internal}
