% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{create_tc}
\alias{create_tc}
\title{Internal function for creation of timecourse plots}
\usage{
create_tc(data, add_CI, colour, quantity = amplitude, mapping = NULL)
}
\arguments{
\item{data}{A data frame to be plotted}

\item{add_CI}{whether to add confidence intervals}

\item{colour}{whether to use colour}

\item{quantity}{The name of the column/quantity to plot}

\item{mapping}{A ggplot2 \code{aes()} mapping.}
}
\description{
Internal function for creation of timecourse plots
}
\keyword{internal}
