% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{cycle_calc}
\alias{cycle_calc}
\title{Calculate cycles}
\usage{
cycle_calc(time_win, frex)
}
\arguments{
\item{time_win}{Time window in seconds.}

\item{frex}{Frequencies of interest.}
}
\value{
the number of cycles for each frequency of interest
}
\description{
A helper function for calculating the appropriate min-max cycles for a fixed
time window/frequency resolution for use with \code{compute_tfr}. For some
analyses you may wish to keep a fixed frequency resolution across the range
being analysed, which requires using a fixed time window. \code{compute_tfr}
expects the minimum and maximum number of cycles to be supplied. Use this
function to calculate the equivalent number of cycles at each frequency.
}
\examples{
cycle_calc(.5, seq(3, 30, length.out = 10))
no_scale_tfr <- compute_tfr(demo_epochs, foi = c(3, 30), n_cycles = range(cycle_calc(0.5, seq(3, 30, length.out = 10))), n_freq = 10)
}
\seealso{
\code{\link{compute_tfr}}
}
