% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{define_fixed_length_events}
\alias{define_fixed_length_events}
\alias{define_fixed_length_events.eeg_data}
\title{Define fixed length events}
\usage{
define_fixed_length_events(data, ...)

\method{define_fixed_length_events}{eeg_data}(data, duration = 1, ...)
}
\arguments{
\item{data}{An \code{eeg_data} object}

\item{...}{Additional arguments.}

\item{duration}{The desired length of the event in seconds. Defaults to 1 sec.}
}
\value{
An eeg_data object with the modified events table
}
\description{
Modify/Add the events table to define fixed length events
}
\section{Methods (by class)}{
\itemize{
\item \code{define_fixed_length_events(eeg_data)}: Modify/Add the events table to
define fixed length events

}}
\seealso{
Other event handlers: 
\code{\link{events}()},
\code{\link{list_epochs}()},
\code{\link{list_events}()},
\code{\link{shift_events}()},
\code{\link{tag_events}()}
}
\concept{event handlers}
