% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_epochs}
\alias{demo_epochs}
\title{A demo \code{eeg_epochs} dataset}
\format{
A \code{eeg_epochs} object
}
\usage{
demo_epochs
}
\description{
This is a small dataset with 11 electrodes and 80 epochs used for
demonstrating some features of the \code{eegUtils} package.
}
\keyword{datasets}
