% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_spatial}
\alias{demo_spatial}
\title{A demo \code{eeg_epochs} dataset}
\format{
An \code{eeg_epochs} object
}
\usage{
demo_spatial
}
\description{
This is a small dataset with 23 electrodes and 100 epochs used for
demonstrating some features of the \code{eegUtils} package. It is part of a
recording from a visual spatial cueing task.
}
\keyword{datasets}
