% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{downsample}
\alias{downsample}
\alias{downsample.eeg_data}
\title{Downsampling EEG data}
\usage{
downsample(data, ...)

\method{downsample}{eeg_data}(data, q, ...)
}
\arguments{
\item{data}{An \code{eeg_data} object to be downsampled}

\item{...}{Parameters passed to functions}

\item{q}{Integer factor to downsample by}
}
\description{
Performs low-pass anti-aliasing filtering and downsamples EEG data by a
specified factor. This is a wrapper for \code{decimate} from the \code{signal} package.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Downsample eeg_data objects
}}

