% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{downsample_events}
\alias{downsample_events}
\title{Downsample the events table}
\usage{
downsample_events(timings, events, srate, q)
}
\arguments{
\item{timings}{the timings from the data}

\item{events}{the events table to downsample}

\item{srate}{sampling rate}

\item{q}{downsampling factor}
}
\description{
Downsample the events table
}
\author{
Matt Craddock \email{matt@craddock.com}
}
\keyword{internal}
