% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{drop_points}
\alias{drop_points}
\title{Drop points before downsampling}
\usage{
drop_points(data, data_length)
}
\arguments{
\item{data}{data to be downsampled}

\item{data_length}{number of points to drop}
}
\description{
Drop points before downsampling
}
\keyword{internal}
