% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{eeg_ar_thresh}
\alias{eeg_ar_thresh}
\alias{eeg_ar_thresh.eeg_data}
\alias{eeg_ar_thresh.eeg_epochs}
\title{Simple thresholding}
\usage{
eeg_ar_thresh(data, threshold, reject = FALSE, ...)

\method{eeg_ar_thresh}{eeg_data}(data, threshold, reject = FALSE, ...)

\method{eeg_ar_thresh}{eeg_epochs}(data, threshold, reject, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_data} or \code{eeg_epochs}.}

\item{threshold}{In microvolts. If one value is supplied, it will be treated as a +- value.}

\item{reject}{If TRUE, remove marked data immediately, otherwise mark for inspection/rejection. Defaults to FALSE.}

\item{...}{Other arguments passed to eeg_ar_thresh}
}
\description{
Reject data based on a simple absolute threshold. By default this will mark datapoints
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Reject data using a simple threshold.

\item \code{eeg_epochs}: Reject data using a simple threshold.
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
