% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{eeg_average.default}
\alias{eeg_average.default}
\title{Default method}
\usage{
\method{eeg_average}{default}(data, ...)
}
\arguments{
\item{data}{An \code{eeg_epochs} object.}

\item{...}{Other arguments passed to the averaging functions}
}
\description{
Default method
}
\author{
Matt craddock \email{matt@mattcraddock.com}
}
