% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eeg_data}
\alias{eeg_data}
\title{Function to create an S3 object of class "eeg_data".}
\usage{
eeg_data(data, srate)
}
\arguments{
\item{data}{Raw data.}

\item{srate}{Sampling rate.}
}
\description{
Function to create an S3 object of class "eeg_data".
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
