% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{eeg_downsample}
\alias{eeg_downsample}
\alias{eeg_downsample.eeg_data}
\alias{eeg_downsample.eeg_epochs}
\title{Downsampling EEG data}
\usage{
eeg_downsample(data, ...)

\method{eeg_downsample}{eeg_data}(data, q, ...)

\method{eeg_downsample}{eeg_epochs}(data, q, ...)
}
\arguments{
\item{data}{An \code{eeg_data} object to be downsampled}

\item{...}{Parameters passed to functions}

\item{q}{Integer factor to downsample by}
}
\description{
Performs low-pass anti-aliasing filtering and downsamples EEG data by a
specified factor. This is a wrapper for \code{decimate} from the
\code{signal} package. Note that this will also adjust the event table,
moving events to the nearest time remaining after downsampling
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_downsample(eeg_data)}: Downsample eeg_data objects

\item \code{eeg_downsample(eeg_epochs)}: Downsample eeg_epochs objects

}}
\examples{
eeg_downsample(demo_epochs, 2)

}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
