% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{eeg_epochs}
\alias{eeg_epochs}
\title{Function to create an S3 object of class "eeg_epochs".}
\usage{
eeg_epochs(data, srate, events = NULL, chan_info = NULL,
  timings = NULL, reference = NULL, epochs = NULL)
}
\arguments{
\item{data}{Raw data - signals from electrodes/channels.}

\item{srate}{Sampling rate in Hz.}

\item{events}{Event table}

\item{chan_info}{String of character names for electrodes.}

\item{timings}{Timing information - samples and sample /sampling rate.}

\item{reference}{Reference channel information, including names of reference
channels, excluded channels etc.}

\item{epochs}{Information about the epochs contained in the data.}
}
\description{
Function to create an S3 object of class "eeg_epochs".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
