% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eeg_evoked}
\alias{eeg_evoked}
\title{Function to create an S3 object of class "eeg_evoked"}
\usage{
eeg_evoked(data, chan_info, timings, ...)
}
\arguments{
\item{data}{evoked data}

\item{chan_info}{Electrode locations etc}

\item{timings}{vector of timepoints}

\item{...}{Other parameters}
}
\description{
Function to create an S3 object of class "eeg_evoked"
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
