% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{eeg_FASTER}
\alias{eeg_FASTER}
\alias{eeg_FASTER.eeg_epochs}
\title{FASTER EEG artefact rejection}
\usage{
eeg_FASTER(.data)

\method{eeg_FASTER}{eeg_epochs}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{eeg_epochs}}

\item{...}{Parameters passed to FASTER}
}
\description{
An implementation of the FASTER artefact rejection method for EEG by Nolan,
Whelan & Reilly (2010) FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods. Not yet fully implemented.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Run FASTER on \code{eeg_epochs}
}}

\references{
Nolan, Whelan & Reilly (2010). FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
