% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{eeg_psd}
\alias{eeg_psd}
\title{Function to create an object of class eeg_psd}
\usage{
eeg_psd(data, srate, events, chan_info = NULL, timings = NULL, freqs,
  dimensions)
}
\arguments{
\item{data}{PSD transformed data}

\item{srate}{Sampling rate in Hz.}

\item{chan_info}{String of character names for electrodes.}

\item{timings}{Timing information - samples and sample /samplirng rate.}

\item{freqs}{vector of frequencies}

\item{dimensions}{List of which dimension is which}

\item{epochs}{Epoch information}
}
\description{
Function to create an object of class eeg_psd
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
