% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_eeg.R
\name{eeg_summarise}
\alias{eeg_summarise}
\alias{eeg_summarise.eeg_epochs}
\title{Calculate simple summary statistics for \verb{eeg_*} objects}
\usage{
eeg_summarise(data, ...)

\method{eeg_summarise}{eeg_epochs}(
  data,
  statistic = c("sem", "mean", "sd", "var"),
  conditions = NULL,
  time_lim = NULL,
  ...
)
}
\arguments{
\item{data}{An \code{eegUtils} object.}

\item{...}{Various arguments passed to specific functions}

\item{statistic}{The statistic to calculate at each timepoint. Defaults to
"sem"}

\item{conditions}{Conditions to group the data by.}

\item{time_lim}{Timepoint(s) to summarise. Can be a range, for which a
summary statistic will be provided for each timepoint, or a list of
individual times. If none is supplied, the function will calculate a
summary for every timepoint.}
}
\value{
A tibble
}
\description{
Calculate the timepoint-by-timepoint mean, standard deviation, standard
error, or variance \code{eeg_epochs} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_summarise(eeg_epochs)}: Calculate summary statistics for \code{eeg_epochs}
objects

}}
