% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{eeg_tfr}
\alias{eeg_tfr}
\title{Object creator for eeg_tfr objects.}
\usage{
eeg_tfr(
  data,
  srate,
  events,
  chan_info = NULL,
  reference,
  timings = NULL,
  freq_info,
  dimensions,
  epochs = NULL
)
}
\arguments{
\item{data}{TFR transformed data}

\item{srate}{Sampling rate in Hz.}

\item{events}{Event tables}

\item{chan_info}{Standard channel information.}

\item{reference}{Reference information}

\item{timings}{Timing information.}

\item{freq_info}{Frequencies and other useful information}

\item{dimensions}{List of which dimension is which}

\item{epochs}{Epoch information}
}
\description{
Object creator for eeg_tfr objects.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
