% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{electrode_locations}
\alias{electrode_locations}
\title{Get standard electrode locations}
\usage{
electrode_locations(data, electrode = "electrode", drop = FALSE,
  plot = FALSE)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{The column name containing electrode names in data.
(Defaults to "electrode").}

\item{drop}{Should electrodes in \code{data} for which default locations
are not available be dropped? (Defaults to FALSE).}

\item{plot}{Plot obtained electrode locations.}
}
\value{
A tibble (or data.frame), or ggplot2 object if \code{plot = TRUE}.
}
\description{
Joins standard electrode locations to EEG data from eegUtils internal data.
}
