% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epoching.R
\name{epoch_data.default}
\alias{epoch_data.default}
\title{Create epochs from EEG data}
\usage{
\method{epoch_data}{default}(data, ...)
}
\arguments{
\item{data}{Continuous data to be epoched.}

\item{...}{Parameters passed to functions}
}
\description{
Create epochs from EEG data
}
