% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{epoch_data.eeg_epochs}
\alias{epoch_data.eeg_epochs}
\title{Create epochs}
\usage{
\method{epoch_data}{eeg_epochs}(data, ...)
}
\arguments{
\item{data}{Continuous data to be epoched}

\item{...}{Parameters passed to methods}
}
\description{
Create epochs
}
