% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{epoch_stats}
\alias{epoch_stats}
\alias{epoch_stats.eeg_epochs}
\title{Epoch statistics}
\usage{
epoch_stats(data, ...)

\method{epoch_stats}{eeg_epochs}(data, ...)
}
\arguments{
\item{data}{Data as a \code{eeg_data} or \code{eeg_epochs} object.}

\item{...}{Other parameters passed to the functions.}
}
\description{
Calculate various statistics for each epoch in the data
}
\section{Methods (by class)}{
\itemize{
\item \code{epoch_stats(eeg_epochs)}: Calculate statistics for each epoch.

}}
\examples{
epoch_stats(demo_epochs)
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
