% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{epochs_tbl}
\alias{epochs_tbl}
\title{Function to create an S3 object of class \code{epochs_tbl}.}
\usage{
epochs_tbl(epochs, recording, epoch_labels, ...)
}
\arguments{
\item{epochs}{epoch column}

\item{recording}{recording name}

\item{epoch_labels}{NULL}

\item{...}{any other parameters to be added}
}
\description{
Function to create an S3 object of class \code{epochs_tbl}.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
