% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_images.R
\name{erp_image}
\alias{erp_image}
\alias{erp_image.default}
\alias{erp_image.eeg_epochs}
\title{Plot ERP images}
\usage{
erp_image(data, electrode = "Cz", smoothing = 10, clim = NULL)

\method{erp_image}{default}(data, electrode = "Cz", smoothing = 10,
  clim = NULL)

\method{erp_image}{eeg_epochs}(data, electrode = "Cz", smoothing = 10,
  clim = NULL)
}
\arguments{
\item{data}{Data frame to be plotted. Requires an amplitude column.}

\item{electrode}{electrode at which to generate an ERP image.}

\item{smoothing}{Number of trials to smooth over when generating image}

\item{clim}{Character vector of min and max values of plotting colour range.
e.g. c(-5,5). Defaults to min and max.}
}
\description{
Plot an ERP image from a single electrode. Smooths over a series of trials in
order to make across-trial patterns more apparent.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default function operates on normal data frames

\item \code{eeg_epochs}: Create an \code{erp_image} from \code{eeg_epochs}
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
