% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_scalp.R
\name{erp_scalp}
\alias{erp_scalp}
\title{Plot ERPs on the scalp}
\usage{
erp_scalp(d, electrode = "electrode", amplitude = "amplitude",
  time = "time", color = NA, size = 0.65, show_guide = TRUE)
}
\arguments{
\item{d}{An EEG dataset.}

\item{electrode}{Column name containing electrode names in data.
Defaults to "electrode".}

\item{amplitude}{Column name containing amplitudes in data.
Defaults to "amplitude".}

\item{time}{Column name containing time in data. Defaults to "time".}

\item{color}{Variable to color lines by. If no variable is passed, only
one line is drawn for each electrode.}

\item{size}{Size of line(s).}

\item{show_guide}{Should a guide be shown.}
}
\value{
Returns a ggplot2 plot object.
}
\description{
Creates an ERP figure for each electrode and layouts them on the scalp.
}
\details{
The function uses default electrode names and locations contained
in the package.
}
\author{
Matti Vuorre, \email{mv2521@columbia.edu}
}
