% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{est_filt_order}
\alias{est_filt_order}
\title{Estimate filter order}
\usage{
est_filt_order(window, tbw, srate)
}
\arguments{
\item{window}{Window to use (string)}

\item{tbw}{Transition bandwidth.}

\item{srate}{Sampling rate (Hz)}
}
\description{
Estimate filter order
}
\keyword{internal}
