% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{est_tbw}
\alias{est_tbw}
\title{Estimate transition bandwidth}
\usage{
est_tbw(filt_pars, srate)
}
\arguments{
\item{filt_pars}{Parsed filter information from parse_filt_freqs}

\item{srate}{Sampling rate (Hz)}
}
\description{
Estimate transition bandwidth
}
\keyword{MNE-python.}
\keyword{Note}
\keyword{are}
\keyword{by}
\keyword{heuristics}
\keyword{internal}
\keyword{same}
\keyword{the}
\keyword{these}
\keyword{used}
