% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{events}
\alias{events}
\alias{events<-}
\alias{events<-.eeg_epochs}
\alias{events<-.eeg_data}
\title{Modify events structure}
\usage{
events(.data)

events(.data) <- value

\method{events}{eeg_epochs}(.data) <- value

\method{events}{eeg_data}(.data) <- value
}
\arguments{
\item{.data}{\code{eegUtils} object to view}

\item{value}{Value to replace `events` structure with.}
}
\description{
Get or set the values in the `events` structure of an eegUtils object.
}
\examples{
events(demo_epochs)
events(demo_epochs) <- mutate(events(demo_epochs),
 sf = dplyr::case_when(
         event_type \%\% 2 == 0 ~ "HSF",
         event_type \%\% 2 == 1 ~ "LSF",
 ))
events(demo_epochs)

}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
