% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_export.R
\name{export_bva}
\alias{export_bva}
\alias{export_bva.eeg_data}
\title{Export continuous data in Brain Vision Analyzer format}
\usage{
export_bva(.data, filename, orientation)

\method{export_bva}{eeg_data}(.data, filename, orientation = "VECTORIZED")
}
\arguments{
\item{.data}{\code{eeg_data} object to be exported.}

\item{filename}{String giving filename to export to. Do NOT include the .vhdr
file extension.}

\item{orientation}{VECTORIZED or MULTIPLEXED. This relates to the way the
data is stored in the binary file. VECTORIZED is the default and
recommended.}
}
\description{
Export continuous EEG data in Brain Vision Analyzer format. This is one of
the recommended formats for BIDS
\url{https://bids-specification.readthedocs.io/en/stable/04-modality-specific-files/03-electroencephalography.html}
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Method for \code{eeg_data}
}}

\keyword{internal}
