% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{faster_chans}
\alias{faster_chans}
\title{Perform global bad channel detection for FASTER}
\usage{
faster_chans(data, sds = 3, ...)
}
\arguments{
\item{data}{A matrix of EEG data signals}

\item{sds}{Standard deviation thresholds}

\item{...}{Further parameters (tbd)}
}
\description{
Perform global bad channel detection for FASTER
}
\keyword{internal}
