% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{faster_cine}
\alias{faster_cine}
\title{FASTER detection of bad channels in single epochs}
\usage{
faster_cine(data, exclude = NULL, ...)
}
\arguments{
\item{data}{\code{eeg_epochs} object.}

\item{exclude}{Channels to be ignored.}

\item{...}{further parameters (tbd)}
}
\description{
FASTER detection of bad channels in single epochs
}
\keyword{internal}
