% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{faster_epochs}
\alias{faster_epochs}
\title{Perform global bad epoch detection for FASTER}
\usage{
faster_epochs(data, sds = 3, ...)
}
\arguments{
\item{data}{\code{eeg_epochs} object}

\item{sds}{standard deviations for threshold}

\item{...}{Further parameters (tbd)}
}
\description{
Perform global bad epoch detection for FASTER
}
\keyword{internal}
