% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{fft_n}
\alias{fft_n}
\title{N-point FFT}
\usage{
fft_n(signal, n)
}
\arguments{
\item{signal}{signal to be FFT'd}

\item{n}{Number of FFT points}
}
\description{
Either zero-pads by adding trailing zeroes, or truncates a signal to N and
runs an FFT.
}
\keyword{internal}
