% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filt_kernel}
\alias{filt_kernel}
\title{Create windowed-sinc filter kernel}
\usage{
filt_kernel(filt_order, cut_off, w)
}
\arguments{
\item{filt_order}{Length of the filter in samples}

\item{cut_off}{Cutoff frequency (normalized as a fraction of sampling rate)}

\item{w}{Window}
}
\description{
Supplied with the length of the filter in samples, the cutoff frequency, and
the windowing function, returns a windowed-sinc filter kernel for use in FIR
filtering.
}
\keyword{internal}
