% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_coefs}
\alias{filter_coefs}
\title{Generate filter coefficients}
\usage{
filter_coefs(method, filt_pars, filter_order, window)
}
\arguments{
\item{method}{IIR or FIR}

\item{filt_pars}{output of parse_filt_freqs}

\item{filter_order}{order of the filter in samples}

\item{window}{Window function name (FIR only)}
}
\description{
Generate filter coefficients for an IIR or FIR filter. FIR coefficients are
generated using \code{\link[signal:fir1]{signal::fir1()}}, which places the -6 dB cutoff precisely at
the specified frequency.
}
\keyword{internal}
