% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{find_times}
\alias{find_times}
\title{Find times in an eeg_* object}
\usage{
find_times(timings, time_lim)
}
\arguments{
\item{timings}{timing information from the EEG object.}

\item{time_lim}{character vector of the time limits}
}
\value{
logical vector of selected timepoints
}
\description{
Internal function to find the rows corresponding to the selected time limits
}
\keyword{internal}
