% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fitting.R
\name{fit_glm}
\alias{fit_glm}
\alias{fit_glm.eeg_epochs}
\title{Fit a GLM model to each timepoint for an individual subject}
\usage{
fit_glm(formula, data, ...)

\method{fit_glm}{eeg_epochs}(formula, data, family = NULL, time_lim = NULL, ...)
}
\arguments{
\item{formula}{Right-hand side A regression formula for a GLM. See ?formula
and notes on use below}

\item{data}{An \code{eegUtils} object.}

\item{...}{Any other arguments passed to (LM/GLM)}

\item{time_lim}{Numeric vector of length 2 specifying time period to be used
as a baseline}
}
\description{
Fits a linear model to each timepoint separately for each electrode.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: GLM fitting for \code{eeg_epochs}
}}

\section{Notes On Usage}{


  The \code{fit_glm} function will fit a linear model to each timepoint for
  each electrode in the input dataset.

  The formula is a standard R formula. Specify only the right-hand side of
  the formula i.e. any predictors.

  The function allows flexible fitting of a baseline covariate, recognising
  the term \code{baseline} in the specified formula.
}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
