% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fitting.R
\name{fit_glm_indiv}
\alias{fit_glm_indiv}
\title{Fit a glm model to each timepoint for an individual subject.}
\usage{
fit_glm_indiv(data, dv, iv_1)
}
\arguments{
\item{data}{An EEG dataset.}

\item{dv}{Column containing dependent variable. (e.g. amplitude)}

\item{iv_1}{Column containing predictor.}
}
\description{
Fit a glm model to each timepoint for an individual subject.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
