% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{gam_topo}
\alias{gam_topo}
\title{Fit a GAM smooth across scalp.}
\usage{
gam_topo(data, scaled_x, scaled_y, grid_res, k = 40)
}
\arguments{
\item{data}{Data frame from which to generate predictions}

\item{scaled_x}{x coordinates rescaled}

\item{scaled_y}{y coordinates rescaled}

\item{grid_res}{resolution of output grid}

\item{k}{number of knots for GAM}
}
\description{
Fit a GAM smooth across scalp.
}
