% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{gauss_filter}
\alias{gauss_filter}
\title{Gaussian filter}
\usage{
gauss_filter(data, srate, freq, fwhm)
}
\arguments{
\item{data}{Data to be filtered}

\item{srate}{Sampling rate of the data}

\item{freq}{Peak frequency of the filter}

\item{fwhm}{Standard deviation of the filter}
}
\description{
Gaussian filtering in the frequency domain.
}
\keyword{internal}
