% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{get_epoch_baselines}
\alias{get_epoch_baselines}
\title{Get epoch baselines}
\usage{
get_epoch_baselines(data, time_lim)
}
\arguments{
\item{data}{data for which to calculate the baselines}

\item{time_lim}{time limits of the baseline period. numeric vector of length
two, c(start, end)}
}
\value{
A numeric matrix of n_epochs x n_channels.
}
\description{
Gets the baseline values for every epoch separately
}
\keyword{internal}
