% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fitting.R
\name{get_lm}
\alias{get_lm}
\title{Calculate coefficients and standard errors}
\usage{
get_lm(mdf, x, inverted_dm, resid_df, robust = FALSE)
}
\arguments{
\item{mdf}{model matrix}

\item{x}{data to be modelled (matrix of trials x channels)}

\item{inverted_dm}{inverted matrix - unscaled covariance}

\item{resid_df}{residual degrees of freedom for the model}

\item{robust}{Use heteroskedasticity-consistent covariance (HC3).}
}
\description{
Calculate coefficients and standard errors
}
\keyword{internal}
