% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_queries.R
\name{get_participant_id}
\alias{get_participant_id}
\alias{get_recording}
\alias{set_participant_id}
\alias{set_recording}
\title{Query and set elements of the \code{epochs} metadata structures}
\usage{
get_participant_id(data)

get_recording(data)

set_participant_id(data, participant_id)

set_recording(data, recording)
}
\arguments{
\item{data}{An \code{eegUtils} object.}

\item{participant_id}{A character vector giving the name to use for the
\code{participant_id} for a particular object.}

\item{recording}{A character vector giving the name to use for the EEG
recording}
}
\value{
\itemize{
\item \code{get_participant_id()} returns a character vector containing the
\code{participant_id}
\item \code{get_recording()} returns a character vector containing
the \code{recording} from values of \code{recording} from the \code{eegUtils} object
\item \code{set_participant_id()} returns the full \code{eegUtils} object with the
\code{participant_id} field modified in the \code{epochs} structure
\item \code{set_recording()} returns the full \code{eegUtils}object with the \code{recording} field modified in the
\code{epochs} structure
}
}
\description{
These functions are used to query and set the \code{participant_id} and
\code{recording} fields of an \code{epochs} structure within any \code{eegUtils} object.
Note that the two \verb{set_*} functions do not operate on \code{eeg_group} objects.
}
\examples{
get_participant_id(demo_epochs)
get_recording(demo_epochs)
set_participant_id(demo_epochs, "002")
set_recording(demo_epochs, "test_recording")
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
