% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{handle_bad_channels}
\alias{handle_bad_channels}
\title{Handle bad channels globally}
\usage{
handle_bad_channels(data, data_chans)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{data_chans}{A vector of channel names to be processed}
}
\value{
The data object with bad channels handled
}
\description{
This function detects and handles globally bad channels using the FASTER algorithm.
}
\keyword{internal}
