% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{handle_bad_channels_in_epochs}
\alias{handle_bad_channels_in_epochs}
\title{Handle bad channels in individual epochs}
\usage{
handle_bad_channels_in_epochs(data, exclude)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{exclude}{Channels to be ignored by FASTER}
}
\value{
The data object with bad channels in individual epochs handled
}
\description{
This function detects and handles bad channels in individual epochs using the FASTER algorithm.
}
\keyword{internal}
