% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{handle_bad_epochs}
\alias{handle_bad_epochs}
\title{Handle bad epochs globally}
\usage{
handle_bad_epochs(data)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}
}
\value{
The data object with bad epochs handled
}
\description{
This function detects and handles globally bad epochs using the FASTER algorithm.
}
\keyword{internal}
