% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{iir_filt}
\alias{iir_filt}
\title{Construct a Butterworth IIR filter.}
\usage{
iir_filt(data, low_freq = NULL, high_freq = NULL, filter_order = 4,
  srate = NULL)
}
\arguments{
\item{data}{Data to be filtered.}

\item{low_freq}{Low passband edge.}

\item{high_freq}{High passband edge.}

\item{filter_order}{Order of the Butterworth filter.}

\item{srate}{Sampling rate of the signal.}
}
\description{
Construct a Butterworth IIR filter.
}
