% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_cnt}
\alias{import_cnt}
\title{Import Neuroscan .CNT file}
\usage{
import_cnt(file_name)
}
\arguments{
\item{file_name}{Name of .CNT file to be loaded.}
}
\description{
Beta version of function to import Neuroscan .CNT files. Only intended for
import of 32-bit files.
}
\keyword{internal}
