% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{import_elc}
\alias{import_elc}
\title{Import ASA .elc electrode location files}
\usage{
import_elc(file_name)
}
\arguments{
\item{file_name}{file name}
}
\description{
Loads and process ASA electrode locations.
ASA electrode locations are given as Cartesian XYZ.
}
\keyword{internal}
