% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{import_elp}
\alias{import_elp}
\title{Import electrode locations from '.elp' file}
\usage{
import_elp(file_name)
}
\arguments{
\item{file_name}{file name of '.elp' electrode locations to import.}
}
\value{
A tibble containing channel locations.
}
\description{
'.elp' files are from BESA.
}
\keyword{internal}
