% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_erplab.R
\name{import_erplab}
\alias{import_erplab}
\title{Import from ERPLAB .erp files}
\usage{
import_erplab(
  file_name,
  df_out = FALSE,
  participant_id = NULL,
  recording = NULL
)
}
\arguments{
\item{file_name}{Filename (and path if not in present working directory)}

\item{df_out}{Defaults to FALSE - outputs an object of class \code{eeg_epochs}. Set to
TRUE to output a data frame.}

\item{participant_id}{By default, the filename will be used as the id of the
participant.}

\item{recording}{By default, the filename will be used as the name of the
recording.}
}
\value{
An object of class \code{eeg_epochs} or a data.frame.
}
\description{
ERPLAB is a toolbox for Event Related Potential analysis associated with
EEGLAB. It exports files in Matlab v6.5 format, either with an .erp or .mat
extension. Note that ERPLAB files can contain data from multiple subjects,
individual subjects, or averaged over multiple subjects. The files do not
distinguish between these possibilities, so users will need to apply their
own knowledge of the contents to import them correctly.
}
\examples{
\dontrun{import_set("your_data.set")}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
