% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_raw}
\alias{import_raw}
\title{Function for reading raw data.}
\usage{
import_raw(file_name, file_path = NULL)
}
\arguments{
\item{file_name}{File to import.}
}
\description{
Currently only BDF or EDF files are supported, and this only provides a wrapper around the \code{edfReader} package.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
