% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_set}
\alias{import_set}
\title{Load \code{EEGLAB} .set files}
\usage{
import_set(
  file_name,
  df_out = FALSE,
  participant_id = NULL,
  recording = NULL,
  drop_custom = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file_name}{Filename (and path if not in present working directory)}

\item{df_out}{Defaults to FALSE - outputs an object of class \code{eeg_data}. Set
to TRUE for a normal data frame.}

\item{participant_id}{By default, the filename will be used as the id of the
participant.}

\item{recording}{By default, the filename will be used as the name of the
recording.}

\item{drop_custom}{Drop custom event fields. TRUE by default.}

\item{verbose}{Print informative messages. TRUE by default.}
}
\value{
An object of class \code{eeg_data}
}
\description{
Load \code{EEGLAB} .set files and convert them to \code{eeg_epochs} objects. Supports
import of files saved both in Matlab v6.5 and Matlab v7.3 formats. Currently,
any ICA weights or decompositions are discarded.
}
\examples{
\dontrun{import_set("your_data.set")}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
