% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{import_tsv}
\alias{import_tsv}
\title{Import ARS '.tsv' electrode location files}
\usage{
import_tsv(file_name)
}
\arguments{
\item{file_name}{file name}
}
\description{
Loads and process ARS electrode locations.
}
\keyword{internal}
