% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{import_txt}
\alias{import_txt}
\title{Import electrode locations from text}
\usage{
import_txt(file_name)
}
\arguments{
\item{file_name}{file name of .txt electrode locations to import.}
}
\value{
A data frame containing standard channel_info
}
\description{
Currently only supports locations given in spherical coordinates.
}
\keyword{internal}
