% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_vhdr}
\alias{import_vhdr}
\title{Function for importing Brain Vision Analyzer files}
\usage{
import_vhdr(file_name, participant_id, recording)
}
\arguments{
\item{file_name}{file name of the header file.}
}
\description{
Function for importing Brain Vision Analyzer files
}
\keyword{internal}
