% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_scalp.R
\name{interactive_scalp}
\alias{interactive_scalp}
\title{Interactive scalp maps}
\usage{
interactive_scalp(data)
}
\description{
Interactive scalp maps
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
