% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{interp_chans}
\alias{interp_chans}
\title{Interpolate channels}
\usage{
interp_chans(.data, bad_chans, missing_coords = FALSE, weights)
}
\arguments{
\item{.data}{Channel data containing all data}

\item{bad_chans}{Vector of names of bad channels}

\item{missing_coords}{Logical vector indicating any channels in the data that
had no associated coordinates}

\item{weights}{Spherical spline weights for interpolation}
}
\description{
Interpolate channels
}
\keyword{internal}
