% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{interp_elecs}
\alias{interp_elecs}
\alias{interp_elecs.eeg_data}
\title{Channel interpolation}
\usage{
interp_elecs(data, bad_elecs, ...)

\method{interp_elecs}{eeg_data}(data, bad_elecs, ...)
}
\arguments{
\item{data}{Data as a \code{eeg_data} or \code{eeg_epochs} object.}

\item{bad_elecs}{Name(s) of electrode(s) to interpolate.}

\item{...}{Other parameters passed to the functions.}
}
\description{
Interpolate EEG channels using a spherical spline (Perrin et al., 1989). The
data must have channel locations attached.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Interpolate EEG channel(s)
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
