% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{is.eeg_data}
\alias{is.eeg_data}
\title{Check if object is of class "eeg_data".}
\usage{
is.eeg_data(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class "eeg_data".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
