% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.eeg_epochs}
\alias{is.eeg_epochs}
\title{Check if object is of class "eeg_epochs".}
\usage{
is.eeg_epochs(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class "eeg_epochs".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
