% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.eeg_erp}
\alias{is.eeg_erp}
\title{Check if object is of class "eeg_erp".}
\usage{
is.eeg_erp(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class "eeg_erp".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
