% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.eeg_evoked}
\alias{is.eeg_evoked}
\title{Check if object is of class \code{eeg_evoked}}
\usage{
is.eeg_evoked(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class \code{eeg_evoked}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
