% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{is.eeg_group}
\alias{is.eeg_group}
\title{Check if object is of class \code{eeg_group}}
\usage{
is.eeg_group(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class \code{eeg_group}
}
