% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{is.eeg_stats}
\alias{is.eeg_stats}
\title{Check if object is of class \code{eeg_stats}}
\usage{
is.eeg_stats(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class \code{eeg_stats}
}
\keyword{internal}
