% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{is.eeg_tfr}
\alias{is.eeg_tfr}
\title{Check if object is of class eeg_tfr}
\usage{
is.eeg_tfr(x)
}
\arguments{
\item{x}{Object to check.}
}
\description{
Check if object is of class eeg_tfr
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
