% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_locs}
\alias{join_locs}
\title{Get standard electrode locations}
\usage{
join_locs(data, ele = "electrode", .drop = FALSE)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{The column name containing electrode names in data.}
}
\description{
Joins standard electrode location information from eegUtils internal data.
}
