% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{joint_diag}
\alias{joint_diag}
\title{Joint diagonalization for SOBI}
\usage{
joint_diag(M, eps, n_channels, pm)
}
\arguments{
\item{M}{Numeric matrix}

\item{eps}{Convergence tolerance}

\item{n_channels}{number of channels}

\item{pm}{number of channels * number of timepoints}
}
\description{
Drop-in replacement for JADE::rjd, which often fails to converge
}
\author{
A. Belouchrani and A. Cichocki. Adapted to R by Matt Craddock
  \email{matt@mattcraddock.com}
}
\keyword{internal}
