% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{legendre_polynomials}
\alias{legendre_polynomials}
\title{Calculate Legendre polynomials up to degree n}
\usage{
legendre_polynomials(n, x)
}
\arguments{
\item{n}{Maximum degree of Legendre polynomials}

\item{x}{Input values}
}
\value{
Matrix of Legendre polynomial values
}
\description{
Calculate Legendre polynomials up to degree n
}
\keyword{internal}
