% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{list_epochs}
\alias{list_epochs}
\title{List epochs}
\usage{
list_epochs(data)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}
}
\description{
List trigger types and any labels found in an \code{eeg_data} object.
}
\seealso{
\code{\link{tag_events}}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
