% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{list_events}
\alias{list_events}
\title{List events}
\usage{
list_events(data)
}
\arguments{
\item{data}{An object of class \code{eeg_data}}
}
\description{
List trigger types and any labels found in an \code{eeg_data} object.
}
\examples{
list_events(demo_epochs)
}
\seealso{
\code{\link{tag_events}} and \code{\link{list_epochs}}

Other event handlers: 
\code{\link{events}()},
\code{\link{list_epochs}()},
\code{\link{tag_events}()}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\concept{event handlers}
