% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\name{make_ear}
\alias{make_ear}
\title{Make an ear shape}
\usage{
make_ear(r, side = c("right", "left"))
}
\arguments{
\item{r}{Radius of head}

\item{side}{"right" or "left"}
}
\description{
Make an ear shape
}
\keyword{internal}
