% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{morlet_res}
\alias{morlet_res}
\title{Morlet wavelet resolutions}
\usage{
morlet_res(frex, n_cycles)
}
\arguments{
\item{frex}{Frequencies of interest}

\item{n_cycles}{Number of cycles for each frequency}
}
\description{
Calculate frequency and temporal standard deviations for the Morlet wavelets
}
\keyword{internal}
