% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{norm_sphere}
\alias{norm_sphere}
\title{Normalize 3d Cartesian co-ordinates to unit sphere}
\usage{
norm_sphere(xyz_coords)
}
\arguments{
\item{xyz_coords}{3D Cartesian electrode locations}
}
\description{
Normalize 3d Cartesian co-ordinates to unit sphere
}
\keyword{internal}
