% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{ortho_norm}
\alias{ortho_norm}
\title{Orthographic electrode projection}
\usage{
ortho_norm(chan_info)
}
\arguments{
\item{chan_info}{Channel information from an eegUtils objects}
}
\value{
A data.frame with x and y columns indictating electrode positions in
  mm
}
\description{
Produce a set of x and y coordinates for plotting from 3D Cartesian
coordinates. This is an orthographic projection of the 3D coordinates,
resulting in bunching up of electrodes at the further reaches of the head.
}
\keyword{internal}
