% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extensions.R
\name{parse_args}
\alias{parse_args}
\title{This function checks which elements of the object need to be
accessed/filtered}
\usage{
parse_args(arg_list, data)
}
\description{
This function checks which elements of the object need to be
accessed/filtered
}
\keyword{internal}
