% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{parse_chaninfo}
\alias{parse_chaninfo}
\title{Parse channel info from an EEGLAB set file}
\usage{
parse_chaninfo(chan_info, drop = FALSE)
}
\arguments{
\item{chan_info}{Channel info list from an EEGLAB set file}

\item{drop}{If there are additional columns, remove all columns except
electrode if TRUE, or just unexpected columns if FALSE.}
}
\description{
Internal function to convert EEGLAB chan_info to eegUtils style
}
\keyword{internal}
