% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{parse_filt_freqs}
\alias{parse_filt_freqs}
\title{Parse filter frequency input}
\usage{
parse_filt_freqs(low_freq, high_freq, srate, method)
}
\arguments{
\item{low_freq}{low frequency cutoff (Hz)}

\item{high_freq}{High frequency cutoff (Hz)}

\item{srate}{Sampling rate (Hz)}

\item{method}{"iir" or "fir" method.}
}
\description{
Parses the frequencies input by the user, converting them to a fraction of
the sampling rate and setting the filter type (low-pass, high-pass,
band-pass, band-stop) appropriately.
}
\keyword{internal}
