% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering2.R
\name{parse_filt_params}
\alias{parse_filt_params}
\title{Parse filter frequency input}
\usage{
parse_filt_params(
  srate,
  method,
  l_freq,
  h_freq,
  l_trans_bw,
  h_trans_bw,
  filter_length,
  window
)
}
\arguments{
\item{srate}{Sampling rate (Hz)}

\item{method}{"iir" or "fir" method.}

\item{l_freq}{low frequency cutoff (Hz)}

\item{h_freq}{High frequency cutoff (Hz)}

\item{l_trans_bw}{Transition bandwidth of high pass (Hz)}

\item{h_trans_bw}{Transition bandwidth of low pass (Hz)}

\item{filter_length}{Filter length of the filter representing the filter order}

\item{window}{Windowing function to use (FIR filtering only). Defaults to
"hamming"; currently only "hann", "hamming" and "blackman" available.}
}
\description{
Parses the frequencies input by the user, converting them to a fraction of
the sampling rate and setting the filter type (low-pass, high-pass,
band-pass, band-stop) appropriately.
}
\keyword{internal}
