% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_butterfly.R
\name{parse_for_bf}
\alias{parse_for_bf}
\title{Parse data for butterfly plots}
\usage{
parse_for_bf(data, time_lim = NULL, baseline = NULL)
}
\arguments{
\item{data}{data to be parsed}

\item{time_lim}{time limits to be returned.}

\item{baseline}{baseline times to be average and subtracted}
}
\description{
Internal command for parsing various data structures into a suitable format
for \code{plot_butterfly}
}
\keyword{internal}
