% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_bva_import.R
\name{parse_vhdr_chans}
\alias{parse_vhdr_chans}
\title{Parse BVA channel info}
\usage{
parse_vhdr_chans(chan_labels, chan_info, verbose = TRUE)
}
\arguments{
\item{chan_labels}{The \verb{Channel Infos} section from a BVA vhdr file}

\item{chan_info}{The \code{Coordinates} section from a BVA vhdr file}
}
\description{
Parse BVA channel info
}
\keyword{internal}
