% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference.R
\name{plot_difference}
\alias{plot_difference}
\alias{plot_difference.eeg_epochs}
\title{Plot ERP difference waves}
\usage{
plot_difference(data, ...)

\method{plot_difference}{eeg_epochs}(
  data,
  electrode = NULL,
  time_lim = NULL,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  conditions = "epoch_labels",
  ...
)
}
\arguments{
\item{data}{\code{eegUtils} object. Should have multiple timepoints.}

\item{...}{Other arguments passed to methods.}

\item{electrode}{Electrode(s) to plot.}

\item{time_lim}{Character vector. Numbers in whatever time unit is used
specifying beginning and end of time-range to plot. e.g. c(-.1, .3)}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean
over the specified period and subtracts. e.g. c(-.1,0)}

\item{colour}{Variable to colour lines by. If no variable is passed, only one
line is drawn.}

\item{color}{Alias for colour.}

\item{mapping}{A ggplot2 \code{aes()} mapping.}

\item{conditions}{Defaults to "epoch_labels".}
}
\value{
Returns a \code{ggplot2} plot object
}
\description{
Calculates the difference between the event-related potentials from two
conditions and plots it.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_difference(eeg_epochs)}: Plot an ERP difference wave from an \code{eeg_epochs} object

}}
\examples{
plot_difference(demo_spatial, conditions = "epoch_labels", electrode = "P8")
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
