% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viewers.R
\name{plot_electrodes}
\alias{plot_electrodes}
\title{Interactive 3d plot of electrode locations}
\usage{
plot_electrodes(data)
}
\arguments{
\item{data}{electrode locations to be plotted in 3d}
}
\description{
Interactive 3d plot of electrode locations
}
