% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gfp.R
\name{plot_gfp}
\alias{plot_gfp}
\title{Plot Global Field Power of EEG Signals}
\usage{
plot_gfp(data, cols = NULL, keep_trials = FALSE)
}
\arguments{
\item{data}{An \code{eeg_epochs} object}

\item{cols}{Condition columns from the \code{epochs} metadata to calculate GFP
separately for different conditions.}

\item{keep_trials}{Calculate GFP for each epoch separately, then average over
epochs.}
}
\description{
Global Field Power (Lehmann & Skrandies, 1980) is a way to quantify the
amount of activity in the overall electroencephalographic signal at a given
timepoint. It corresponds to the spatial standard deviation.
}
\examples{
plot_gfp(demo_spatial)
plot_gfp(demo_spatial, keep_trials = TRUE)
plot_gfp(demo_spatial, cols = "epoch_labels")
plot_gfp(demo_spatial, cols = "epoch_labels", keep_trials = TRUE)
}
