% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plotting.R
\name{plot_psd}
\alias{plot_psd}
\alias{plot_psd.eeg_epochs}
\alias{plot_psd.eeg_data}
\alias{plot_psd.eeg_ICA}
\alias{plot_psd.data.frame}
\title{Plot power spectral density}
\usage{
plot_psd(data, freq_range = NULL, ...)

\method{plot_psd}{eeg_epochs}(data, freq_range = NULL, ...)

\method{plot_psd}{eeg_data}(data, freq_range = NULL, ...)

\method{plot_psd}{eeg_ICA}(data, freq_range = NULL, components = NULL,
  ...)

\method{plot_psd}{data.frame}(data, freq_range = NULL, ...)
}
\arguments{
\item{data}{Object of class \code{eeg_epochs}}

\item{freq_range}{Vector of lower and upper frequencies to plot. (e.g. c(1,
40))}

\item{...}{Additional parameters}

\item{components}{Which components to compute the PSD for}
}
\description{
Calculate and plot the PSD for \code{eeg_*} objects. Output units are dB. 512
points are used for FFTs.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Plot PSDs for \code{eeg_epochs}.

\item \code{eeg_data}: Plot PSDs for \code{eeg_epochs}.

\item \code{eeg_ICA}: Plot PSD for \code{eeg_ICA} objects

\item \code{data.frame}: Plot PSDs for \code{data.frame}s.
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
