% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plotting.R
\name{plot_tfr}
\alias{plot_tfr}
\title{Time-frequency plot}
\usage{
plot_tfr(
  data,
  electrode = NULL,
  time_lim = NULL,
  freq_range = NULL,
  baseline_type = "none",
  baseline = NULL,
  fill_lims = NULL,
  interpolate = FALSE
)
}
\arguments{
\item{data}{Object of class \code{eeg_tfr}}

\item{electrode}{Electrode to plot. If none is supplied, averages over all
electrodes.}

\item{time_lim}{Time limits of plot.}

\item{freq_range}{Vector of two numbers. (e.g. c(8, 40)).}

\item{baseline_type}{baseline correction to apply. Defaults to "none".}

\item{baseline}{Baseline period}

\item{fill_lims}{Custom colour scale (i.e. range of power). e.g. c(-5, 5).}

\item{interpolate}{Interpolation of raster for smoother plotting.}
}
\description{
Creates a time-frequency plot of an \code{eeg_tfr} object. The plot has time
on the x-axis and frequency on the y-axis. If no electrode is supplied, it
will average over all electrodes.
}
\details{
Various different baseline options can be applied here (e.g. "db" for
decibels, "pc" for percent change, "divide" for division; see
\code{rm_baseline} for details).
}
\seealso{
[rm_baseline()]
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
