% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{plot_timecourse}
\alias{plot_timecourse}
\title{Plot 1d timecourse data.}
\usage{
plot_timecourse(df, time_lim = NULL, group = NULL, facet = NULL,
  add_CI = FALSE, baseline = NULL)
}
\arguments{
\item{df}{EEG dataset. Should have multiple timepoints.}

\item{time_lim}{Character vector. Numbers in whatever time unit is used specifying beginning and end of time-range to plot. e.g. c(-100,300)}

\item{facet}{Create multiple plots for a specified grouping variable.}

\item{add_CI}{Add confidence intervals to the graph. Defaults to 95 percent between-subject CIs.}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean over the specified period and subtracts. e.g. c(-100,0)}
}
\value{
Returns a ggplot2 plot object
}
\description{
Typically event-related potentials/fields, but could also be timecourses from frequency analyses for single frequencies. Averages over all submitted electrodes. Output is a ggplot2 object.
}
