% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{pol_to_sph}
\alias{pol_to_sph}
\title{Convert EEGLAB polar to spherical coordinates}
\usage{
pol_to_sph(theta, phi)
}
\arguments{
\item{theta}{Azimuth from polar co-ordinates (theta in supplied electrode
locations)}

\item{phi}{Elevation from polar co-ordinates (radius in supplied electrode
locations)}
}
\description{
Hard-coded to a radius of 85 mm (as in BESA).
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
