% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for FASTER processing}
\usage{
prepare_data(data, exclude)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{exclude}{Channels to be ignored by FASTER}
}
\value{
A list containing prepared data information
}
\description{
This function prepares the data for processing, handling exclusions and storing original reference information.
}
\keyword{internal}
