% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_data}
\alias{print.eeg_data}
\title{Print eeg_data summary}
\usage{
\method{print}{eeg_data}(x, ...)
}
\arguments{
\item{x}{\code{eeg_data} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_data} object
}
