% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_group}
\alias{print.eeg_group}
\title{Print \code{eeg_group} summary}
\usage{
\method{print}{eeg_group}(x, ...)
}
\arguments{
\item{x}{\code{eeg_group} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_group} object
}
