% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_ICA}
\alias{print.eeg_ICA}
\title{Print eeg_epochs summary}
\usage{
\method{print}{eeg_ICA}(x, ...)
}
\arguments{
\item{x}{\code{eeg_epochs} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_epochs} object
}
