% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_lm}
\alias{print.eeg_lm}
\title{Print \code{eeg_lm} summary}
\usage{
\method{print}{eeg_lm}(x, ...)
}
\arguments{
\item{x}{\code{eeg_lm} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_lm} object
}
