% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_stats}
\alias{print.eeg_stats}
\title{Print eeg_stats summary}
\usage{
\method{print}{eeg_stats}(x, ...)
}
\arguments{
\item{x}{\code{eeg_stats} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_stats} object
}
