% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.eeg_tfr}
\alias{print.eeg_tfr}
\title{Print \code{eeg_tfr} summary}
\usage{
\method{print}{eeg_tfr}(x, ...)
}
\arguments{
\item{x}{\code{eeg_tfr} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_tfr} object
}
