% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print.tfr_average}
\alias{print.tfr_average}
\title{Print \code{tfr_average} summary}
\usage{
\method{print}{tfr_average}(x, ...)
}
\arguments{
\item{x}{\code{tfr_average} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of a \code{tfr_average} object
}
