% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{proc_events}
\alias{proc_events}
\title{Internal function for processing epoch_events during selection}
\usage{
proc_events(epoch_events, event_type, epoch_nos, event_labels, keep)
}
\description{
Converts character strings into a vector of epoch numbers with matching labels.
}
\keyword{internal}
