% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{project_elecs}
\alias{project_elecs}
\title{Electrode projection}
\usage{
project_elecs(chan_info, method = "stereographic")
}
\arguments{
\item{chan_info}{Channel information from an eegUtils object}

\item{method}{Method of projection. "stereographic" or "orthographic".
Defaults to sterographic.}
}
\description{
Project a set of 3D Cartesian co-ordinates to a 2D plane for plotting. The
projection can be orthographic or stereographic. Orthographic is closer to
how the scalp would look from above, since it does not compensate for
height/distance from the xy plane. This causes bunching up of electrodes near
the limits of the head.Stereographic preserves more of the general shape of
the features by "unrolling" the electrode positions.
}
\keyword{internal}
