% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_dat}
\alias{read_dat}
\title{Read the raw data for a BVA file.}
\usage{
read_dat(file_name, file_size, bin_format, multiplexed)
}
\arguments{
\item{file_name}{Filename of the .dat file}

\item{file_size}{Size of the .dat file}

\item{bin_format}{Storage format.}

\item{multiplexed}{Logical; whether the data is VECTORIZED (FALSE) or MULTIPLEXED (TRUE)}
}
\description{
Read the raw data for a BVA file.
}
\keyword{internal}
