% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{read_vmrk}
\alias{read_vmrk}
\title{Read BVA markers}
\usage{
read_vmrk(file_name)
}
\arguments{
\item{file_name}{File name of the .vmrk markers file.}
}
\description{
Import and parse BVA marker files.
}
\keyword{internal}
