% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{rearrange_tfr}
\alias{rearrange_tfr}
\title{Rearrange and combine tfr_average objects}
\usage{
rearrange_tfr(data, all_l)
}
\description{
Rearrange and combine tfr_average objects
}
\keyword{internal}
