% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{remove_break_periods}
\alias{remove_break_periods}
\alias{remove_break_periods.eeg_data}
\title{Remove break periods from raw EEG data}
\usage{
remove_break_periods(data, ...)

\method{remove_break_periods}{eeg_data}(
  data,
  break_threshold = 2,
  start_offset = 1.5,
  end_offset = 0.5,
  ...
)
}
\arguments{
\item{data}{An \code{eeg_data} object from which break periods are to be removed}

\item{...}{Other parameters for future releases}

\item{break_threshold}{Minimum amount of time without an event code to be
considered as the break period. Defaults to 2 sec.}

\item{start_offset}{Offset to add to the beginning time of the break period.
Defaults to 1.5 sec.}

\item{end_offset}{Offset to subtract from the end of the break period}
}
\description{
Detects and removes data related to the break periods based on threshold
specified by the user. The user can also change the offset values used to
avoid clipping of data for future epochs. Make sure threshold and offset
values are provided in seconds.
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_break_periods(eeg_data)}: Remove break periods from raw EEG data

}}
\author{
Akash Mer
}
