% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reref_}
\alias{reref_}
\title{Referencing.}
\usage{
reref_(data, ref_chans = "all")
}
\arguments{
\item{data}{Data to re-reference.}

\item{ref_chans}{Channels to reference data to. Defaults to "all" i.e. average of all electrodes in data.}
}
\description{
Referencing.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
