% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{restore_original_reference}
\alias{restore_original_reference}
\title{Restore original reference}
\usage{
restore_original_reference(data, metadata)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{metadata}{A list containing metadata information, including the original reference}
}
\value{
The data object with the original reference restored
}
\description{
This function restores the original reference if it existed.
}
\keyword{internal}
