% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencing.R
\name{rm_baseline}
\alias{rm_baseline}
\title{Baseline correction.}
\usage{
rm_baseline(data, time_lim = NULL)
}
\arguments{
\item{data}{Data to be baseline corrected.}

\item{time_lim}{Numeric character vector (e.g. time_lim <- c(-.1, 0)). If none given, defaults to mean of whole epoch if the data is epoched, or the channel mean if the data is continuous.}
}
\description{
Baseline correction.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
