% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{rotate_sph}
\alias{rotate_sph}
\title{Rotate spherical coordinates and recalculate others}
\usage{
rotate_sph(chan_info, degrees)
}
\arguments{
\item{chan_info}{channel information structure}

\item{degrees}{degrees by which to rotate elecs}
}
\description{
Rotate spherical coordinates and recalculate others
}
\keyword{internal}
