% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{run_fir}
\alias{run_fir}
\title{Run FIR filter using overlap-add FFT}
\usage{
run_fir(.data, filt_coef, filter_order)
}
\arguments{
\item{.data}{Data to be filtered.}

\item{filt_coef}{Filter coefficients}

\item{filter_order}{Order of filter}
}
\description{
Run FIR filter using overlap-add FFT
}
\keyword{internal}
