% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{samples}
\alias{samples}
\title{Get number of samples}
\usage{
samples(.data)
}
\arguments{
\item{.data}{Object to get total number of sampling points from}
}
\description{
Get number of samples
}
\keyword{internal}
