% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_elecs.default}
\alias{select_elecs.default}
\title{Select electrodes from a given dataset.}
\usage{
\method{select_elecs}{default}(data, electrode = NULL, keep = TRUE, ...)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{A character vector of electrode labels for selection or removal.}

\item{keep}{Defaults to TRUE. Set to false to *remove* the selected electrodes.}

\item{...}{Arguments used with related methods}
}
\value{
Data frame with only data from the chosen electrodes
}
\description{
Checks for presence of electrode column, and if found, presence of selected electrodes.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
