% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_elecs.eeg_data}
\alias{select_elecs.eeg_data}
\title{Select electrodes from a given dataset.}
\usage{
\method{select_elecs}{eeg_data}(data, electrode, keep = TRUE,
  df_out = FALSE)
}
\arguments{
\item{data}{An \code{eeg_data} object.}

\item{electrode}{A character vector of electrode labels for selection or removal.}

\item{keep}{Defaults to TRUE. Set to false to *remove* the selected electrodes.}

\item{df_out}{}
}
\value{
\code{eeg_data} object with selected electrodes removed/kept.
}
\description{
Checks for presence of electrode column, and if found, presence of selected electrodes.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
