% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_elecs}
\alias{select_elecs}
\title{Select electrodes from a given dataset.}
\usage{
select_elecs(data, electrode, keep = TRUE)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{A character vector of electrode labels for selection or removal.}

\item{keep}{Defaults to TRUE. Set to false to *remove* the selected electrodes.}
}
\value{
Data frame with only data from the chosen electrodes
}
\description{
Checks for presence of electrode column, and if found, presence of selected electrodes.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
