% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_epochs.default}
\alias{select_epochs.default}
\title{Select epochs from eeg_data}
\usage{
\method{select_epochs}{default}(data, ...)
}
\arguments{
\item{data}{\code{eeg_data} object from which to select epochs.}

\item{...}{Parameters passed to specific functions}
}
\description{
Select epochs from eeg_data
}
