% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_epochs.eeg_data}
\alias{select_epochs.eeg_data}
\title{Select epochs from \code{eeg_data}}
\usage{
\method{select_epochs}{eeg_data}(data, epoch_no = NULL, keep = TRUE,
  df_out = FALSE, ...)
}
\arguments{
\item{data}{\code{eeg_data} object from which to select epochs.}

\item{epoch_no}{Epoch numbers to select}

\item{keep}{Defaults to TRUE, meaning select the specified epochs. Set to FALSE to remove specified epochs.}

\item{df_out}{Output a data.frame instead of an eeg_data object.}

\item{...}{Parameters passed to specific functions}
}
\description{
Select epochs from \code{eeg_data}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
