% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_freqs}
\alias{select_freqs}
\alias{select_freqs.eeg_tfr}
\title{Select frequencies}
\usage{
select_freqs(data, freq_range)

\method{select_freqs}{eeg_tfr}(data, freq_range)
}
\arguments{
\item{data}{An \code{eeg_tfr} object.}

\item{freq_range}{The range of frequencies to retain. Can be a scale or the
lower and upper bounds. (e.g. c(5, 30))}
}
\description{
Select specific frequencies from \code{eeg_tfr} objects. Can be used to
selecting either single frequencies or anything within a range.
}
\section{Methods (by class)}{
\itemize{
\item \code{select_freqs(eeg_tfr)}: Function for selecting specific frequencies from \code{eeg_tfr} objects.

}}
\examples{
demo_tfr <- compute_tfr(demo_epochs, foi = c(4, 30), n_freq = 10, n_cycles = 5)
select_freqs(demo_tfr, c(8, 12))
}
