% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_times.default}
\alias{select_times.default}
\title{Select timerange}
\usage{
\method{select_times}{default}(data, ..., time_lim)
}
\arguments{
\item{data}{An EEG dataset in a data frame. Must have a column named "time".}

\item{...}{Arguments used with related methods}

\item{time_lim}{A character vector of two numbers indicating the time range to be selected e.g. c(min, max)}
}
\value{
Data frame with only data from within the specified range.
}
\description{
Select timerange from a dataframe.
}
\seealso{
\code{\link{select_times}}, \code{\link{select_times.eeg_data}}, \code{\link{select_elecs}}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
