% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_times.eeg_data}
\alias{select_times.eeg_data}
\title{Select time range}
\usage{
\method{select_times}{eeg_data}(data, time_lim, df_out = FALSE)
}
\arguments{
\item{data}{Must be an object of class \code{eeg_data}.}

\item{time_lim}{A character vector of two numbers indicating the time range
to be selected e.g. c(min, max)}

\item{df_out}{Return a data frame rather than an \code{eeg_data} object.
Defaults to FALSE (i.e. returns an \code{eeg_data} object)}
}
\description{
select a specifed time range from objects of class \code{eeg_data}.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
