% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_times}
\alias{select_times}
\title{Select timepoints from a given dataset}
\usage{
select_times(data, time_lim)
}
\arguments{
\item{data}{An EEG dataset.}

\item{time_lim}{A character vector of two numbers indicating the time range to be selected e.g. c(min, max)}
}
\value{
Data frame with only data from within the specified range.
}
\description{
Select timepoints from a given dataset
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
