% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{set_cmap}
\alias{set_cmap}
\title{Set colourmap and limits for topoplot}
\usage{
set_cmap(topo, colourmap, clim = NULL)
}
\arguments{
\item{topo}{ggplot2 object produced by topoplot command}

\item{colourmap}{Requested colourmap}

\item{clim}{Limits of colour scale}
}
\description{
Set colourmap and limits for topoplot
}
