% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{set_palette}
\alias{set_palette}
\title{Set palette and limits for topoplot}
\usage{
set_palette(topo, palette, limits = NULL)
}
\arguments{
\item{topo}{ggplot2 object produced by topoplot command}

\item{palette}{Requested palette}

\item{limits}{Limits of colour scale}
}
\description{
Set palette and limits for topoplot
}
\keyword{internal}
