% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{shift_events}
\alias{shift_events}
\alias{shift_events.eeg_data}
\title{Shift events by a fixed amount of time}
\usage{
shift_events(data, event_codes, delay, ...)

\method{shift_events}{eeg_data}(data, event_codes, delay, ...)
}
\arguments{
\item{data}{An \code{eeg_data} object}

\item{event_codes}{Numerical event codes for which the delay has to be
applied}

\item{delay}{Time delay in seconds}

\item{...}{Additional arguments.}
}
\value{
An eeg_data object with the specified events corrected
}
\description{
Shift event onset and timings to account for monitor/display device delay
measured by a photo sensor
}
\section{Methods (by class)}{
\itemize{
\item \code{shift_events(eeg_data)}: Shift events in an \code{eeg_data} object.

}}
\seealso{
Other event handlers: 
\code{\link{events}()},
\code{\link{list_epochs}()},
\code{\link{list_events}()},
\code{\link{tag_events}()}
}
\concept{event handlers}
