% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{sobi_ICA}
\alias{sobi_ICA}
\title{SOBI ICA}
\usage{
sobi_ICA(data)
}
\arguments{
\item{data}{Data to be ICAed.}
}
\description{
Internal function for running SOBI ICA on an \code{eeg_epochs} object
}
\keyword{internal}
