% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{spec_inv}
\alias{spec_inv}
\title{Spectral inversion}
\usage{
spec_inv(filt_kern)
}
\arguments{
\item{filt_kern}{Filter kernel to be inverted}
}
\description{
Convert high-pass to low-pass, band-pass to band-stop, and vice versa.
}
\keyword{internal}
