% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{sph_to_topo}
\alias{sph_to_topo}
\title{Convert spherical to topographical co-ordinates}
\usage{
sph_to_topo(theta, phi)
}
\arguments{
\item{theta}{Theta}

\item{phi}{Phi}
}
\description{
Expects input in degrees
}
\keyword{internal}
