% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{spheric_spline}
\alias{spheric_spline}
\title{Calculate a spherical spline smooth for interpolation of electrodes}
\usage{
spheric_spline(good_elecs, bad_elecs, data)
}
\arguments{
\item{good_elecs}{Electrodes with positions that do not need interpolation}

\item{bad_elecs}{Electrodes to be interpolated}

\item{data}{Raw data}
}
\description{
Calculate a spherical spline smooth for interpolation of electrodes
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
