% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{split_vec}
\alias{split_vec}
\title{Segment data}
\usage{
split_vec(vec, seg_length, overlap, detrend = "mean")
}
\arguments{
\item{vec}{Data vector to be split up into segments.}

\item{seg_length}{Length of segments to be FFT'd (in samples).}

\item{overlap}{Overlap between segments (in samples).}

\item{detrend}{Detrend segments. Defaults to "mean" - removes mean from each
segment. Anything else turns off detrending.}
}
\description{
Split data into segments for Welch PSD. Any leftover data is discarded (i.e.
if seg_length is 256 and signal length is 400, only 1 segment is returned)
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
