% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\name{stat_scalpmap}
\alias{stat_scalpmap}
\alias{geom_head}
\alias{geom_mask}
\alias{geom_ears}
\alias{geom_channels}
\title{Create an interpolated scalp surface}
\usage{
stat_scalpmap(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  grid_res = 200,
  interpolate = FALSE,
  interp_limit = c("skirt", "head"),
  method = "biharmonic",
  r = NULL,
  ...
)

geom_head(
  mapping = NULL,
  data = NULL,
  show.legend = NA,
  na.rm = TRUE,
  inherit.aes = TRUE,
  interp_limit = "skirt",
  r = 95,
  ...
)

geom_mask(
  mapping = NULL,
  data = NULL,
  show.legend = NA,
  na.rm = FALSE,
  colour = "white",
  size = rel(5),
  linewidth = rel(5),
  r = 95,
  interp_limit = "skirt",
  ...
)

geom_ears(
  mapping = NULL,
  data = NULL,
  show.legend = NA,
  na.rm = FALSE,
  r = 95,
  ...
)

geom_channels(
  mapping = NULL,
  data = NULL,
  geom = "point",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{grid_res}{Resolution of the interpolation grid. (Defaults to 200
points).}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{interp_limit}{Topoplot with a "skirt" or inside the "head".}

\item{method}{"biharmonic" or "gam"}

\item{r}{Radius of head}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{colour}{For \code{geom_mask}, colour of the masking ring.}

\item{size}{For \code{geom_mask}, width of the masking ring. Deprecated.}

\item{linewidth}{For \code{geom_mask}, width of the masking ring.}

\item{geom}{"point" for points or "text" for labels. Default is "point".}
}
\description{
\code{stat_scalpmap} creates an interpolated surface for an irregular set of
x-y coordinates, as is typically required for a topographical EEG plot. Since
the surface should be approximately round, the function attempts to blank out
portions of the surface that lay outside the area within the electrodes.

\code{geom_head()} adds a headshape to a plot.

\code{geom_mask()} adds a masking ring to smooth the edges of a scalp map
generated by \code{stat_scalpmap()}, to give it a circular appearance.

\code{geom_ears} simply draws a pair of ears attached to the head shape.

\code{geom_channels} adds either points or text labels at channel locations.
This is a convenience function to prevent overplotting when the input data
contains many rows of data.
}
\seealso{
Other topoplot functions: 
\code{\link{geom_topo}()},
\code{\link{stat_scalpcontours}()}

Other topoplot functions: 
\code{\link{geom_topo}()},
\code{\link{stat_scalpcontours}()}

Other topoplot functions: 
\code{\link{geom_topo}()},
\code{\link{stat_scalpcontours}()}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\concept{topoplot functions}
