% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\docType{data}
\name{StatBiharmonic}
\alias{StatBiharmonic}
\title{StatBiharmonic}
\format{An object of class \code{StatBiharmonic} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatBiharmonic
}
\description{
StatBiharmonic
}
\keyword{datasets}
