% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_contour_test.R
\docType{data}
\name{StatScalpContours}
\alias{StatScalpContours}
\title{StatScalpcontours}
\format{
An object of class \code{StatScalpContours} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatScalpContours
}
\description{
StatScalpcontours
}
\keyword{internal}
