% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\docType{data}
\name{StatScalpmap}
\alias{StatScalpmap}
\title{StatScalpmap}
\format{An object of class \code{StatScalpmap} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatScalpmap
}
\description{
StatScalpmap
}
\keyword{internal}
