% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{summary.eeg_data}
\alias{summary.eeg_data}
\title{Print eeg_data summary}
\usage{
\method{summary}{eeg_data}(object, ...)
}
\arguments{
\item{object}{\code{eeg_data} object to be printed}

\item{...}{Further arguments passed}
}
\description{
Print a basic summary of the contents of an \code{eeg_data} object
}
