% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{tag_epochs}
\alias{tag_epochs}
\alias{tag_epochs.default}
\alias{tag_epochs.eeg_epochs}
\title{Tag epochs with labels}
\usage{
tag_epochs(.data, ...)

\method{tag_epochs}{default}(.data, ...)

\method{tag_epochs}{eeg_epochs}(.data, event_type = NULL, event_label = NULL, ...)
}
\arguments{
\item{.data}{An \code{eegUtils} object}

\item{...}{Additional arguments.}

\item{event_type}{Label epochs according to specific event_types (typically a
trigger)}

\item{event_label}{Label epochs according to specific event_labels}
}
\description{
Tag epochs with labels indicating details such as experimental condition,
based on the occurrence of event triggers from the events() structure. This
adds a new column to the epochs structure in an \code{eeg_epochs} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{tag_epochs(eeg_epochs)}: Tag epochs in an \code{eeg_epochs} object.

}}
