% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{tag_events}
\alias{tag_events}
\title{Tag events}
\usage{
tag_events(data)
}
\arguments{
\item{data}{}
}
\description{
Give trigger events meaningful labels.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
