% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{tf_morlet}
\alias{tf_morlet}
\title{Perform Morlet time-frequency analysis}
\usage{
tf_morlet(data, foi, n_freq, n_cycles, keep_trials, output)
}
\arguments{
\item{data}{Data in \code{eeg_epochs} format.}

\item{foi}{Frequencies of interest. Scalar or character vector of the lowest
and highest frequency to resolve.}

\item{n_freq}{Number of frequencies to be resolved.}

\item{n_cycles}{Number of cycles at each frequency.}

\item{keep_trials}{Keep single trials or average over them before returning.}

\item{output}{Sets whether output is power, phase, or fourier coefficients.}
}
\description{
Perform Morlet time-frequency analysis
}
\keyword{internal}
