% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{topo_norm}
\alias{topo_norm}
\title{Convert topographical 2d to cartesian 2d}
\usage{
topo_norm(angle, radius)
}
\arguments{
\item{angle}{Angle}

\item{radius}{Radius}
}
\description{
Expects input to be in degrees
}
\keyword{internal}
