% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot.default}
\alias{topoplot.default}
\title{Topographical Plotting Function for EEG}
\usage{
\method{topoplot}{default}(data, ...)
}
\arguments{
\item{data}{An object passed to the function}

\item{...}{Any other parameters}
}
\description{
Topographical Plotting Function for EEG
}
