% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot}
\alias{topoplot}
\alias{topoplot.default}
\alias{topoplot.data.frame}
\alias{topoplot.eeg_data}
\alias{topoplot.eeg_epochs}
\alias{topoplot.eeg_ICA}
\alias{topoplot.eeg_tfr}
\title{Topographical Plotting Function for EEG}
\usage{
topoplot(data, ...)

\method{topoplot}{default}(data, ...)

\method{topoplot}{data.frame}(
  data,
  time_lim = NULL,
  limits = NULL,
  chanLocs = NULL,
  method = "biharmonic",
  r = NULL,
  grid_res = 200,
  palette = "RdBu",
  interp_limit = "skirt",
  contour = TRUE,
  chan_marker = "point",
  quantity = "amplitude",
  montage = NULL,
  highlights = NULL,
  scaling = 1,
  groups = NULL,
  verbose = TRUE,
  k = 40,
  fill_title = expression(paste("Amplitude (", mu, "V)")),
  ...
)

\method{topoplot}{eeg_data}(
  data,
  time_lim = NULL,
  limits = NULL,
  chanLocs = NULL,
  method = "biharmonic",
  r = NULL,
  grid_res = 200,
  palette = "RdBu",
  interp_limit = "skirt",
  contour = TRUE,
  chan_marker = "point",
  quantity = "amplitude",
  montage = NULL,
  highlights = NULL,
  scaling = 1,
  verbose = TRUE,
  groups = NULL,
  k = 40,
  fill_title = expression(paste("Amplitude (", mu, "V)")),
  ...
)

\method{topoplot}{eeg_epochs}(
  data,
  time_lim = NULL,
  limits = NULL,
  chanLocs = NULL,
  method = "biharmonic",
  r = NULL,
  grid_res = 200,
  palette = "RdBu",
  interp_limit = "skirt",
  contour = TRUE,
  chan_marker = "point",
  quantity = "amplitude",
  montage = NULL,
  highlights = NULL,
  scaling = 1,
  groups = NULL,
  verbose = TRUE,
  k = 40,
  fill_title = expression(paste("Amplitude (", mu, "V)")),
  ...
)

\method{topoplot}{eeg_ICA}(
  data,
  component,
  time_lim = NULL,
  limits = NULL,
  chanLocs = NULL,
  method = "biharmonic",
  r = NULL,
  grid_res = 200,
  palette = "RdBu",
  interp_limit = "skirt",
  contour = TRUE,
  chan_marker = "point",
  quantity = "amplitude",
  montage = NULL,
  highlights = NULL,
  scaling = 1,
  verbose = TRUE,
  groups = NULL,
  k = 40,
  fill_title = "a.u.",
  ...
)

\method{topoplot}{eeg_tfr}(
  data,
  time_lim = NULL,
  limits = NULL,
  chanLocs = NULL,
  method = "biharmonic",
  r = NULL,
  grid_res = 200,
  palette = "RdBu",
  interp_limit = "skirt",
  contour = TRUE,
  chan_marker = "point",
  quantity = "power",
  montage = NULL,
  highlights = NULL,
  scaling = 1,
  freq_range = NULL,
  verbose = TRUE,
  groups = NULL,
  k = 40,
  fill_title = expression(paste("Power (", mu, V^2, ")")),
  ...
)
}
\arguments{
\item{data}{An EEG dataset. If the input is a data.frame, then it must have
columns \code{electrode}, \code{x}, \code{y}, and \code{amplitude}. \code{x} and \code{y} should be
(Cartesian) electrode co-ordinates.}

\item{...}{Various arguments passed to specific functions}

\item{time_lim}{Timepoint(s) to plot. Can be one time, a range to average
over, or a list of individual times. If none is supplied, the function will
average across all timepoints in the supplied data.}

\item{limits}{Limits of the fill scale - should be given as a character
vector with two values specifying the start and endpoints e.g. limits =
c(-2,-2). Will ignore anything else. Defaults to the range of the data.}

\item{chanLocs}{Allows passing of channel locations (see
\code{electrode_locations})}

\item{method}{Interpolation method. "Biharmonic" or "gam". "Biharmonic"
implements the same method used in Matlab's EEGLAB. "gam" fits a
Generalized Additive Model with k = 40 knots. Defaults to biharmonic spline
interpolation.}

\item{r}{Radius of cartoon head_shape in mm. Default value is 95 (mm) when
using \code{interp_limit = "skirt"}.}

\item{grid_res}{Resolution of the interpolated grid. Higher = smoother but
slower. Defaults to 200 points per edge.}

\item{palette}{Defaults to RdBu if none supplied. Can be any palette from
RColorBrewer or viridis. If an unsupported palette is specified, switches
to Greens.}

\item{interp_limit}{"skirt", "head", or "convex_hull". Defaults to "skirt".
"skirt" interpolates just past the farthest electrode and does not respect
the boundary of the head_shape. "head" interpolates up to the radius of the
plotted head. "convex_hull" clips the interpolated surface to the convex
hull of the electrode positions, preventing extrapolation beyond the
electrode coverage area.}

\item{contour}{Plot contour lines on topography (defaults to TRUE)}

\item{chan_marker}{Set marker for electrode locations. "point" = point,
"name" = electrode name, "none" = no marker. Defaults to "point".}

\item{quantity}{Allows plotting of an arbitrary quantitative column. Defaults
to amplitude. Use quoted column names. E.g. "p.value", "t_statistic".}

\item{montage}{Name of an existing montage set. Defaults to NULL. (currently
only 'biosemi64alpha' available other than default 10/20 system)}

\item{highlights}{Electrodes to highlight (in white).}

\item{scaling}{Scaling multiplication factor for labels and any plot lines.
Defaults to 1.}

\item{groups}{Column name for groups to retain. This is required to create
facetted plots.}

\item{verbose}{Warning messages when electrodes do not have locations.
Defaults to TRUE.}

\item{k}{Degrees of freedom used for spline when using \code{method = gam}.
Defaults to 40.}

\item{fill_title}{Allows specification of the title of the colourbar.
Defaults to `expression(paste("Amplitude (", mu, "V)")), appropriate for
typical time-domain data.}

\item{component}{Component to plot (numeric)}

\item{freq_range}{Range of frequencies to average over.}
}
\description{
Allows topographical plotting of functional data. Output is a ggplot2 object.
}
\section{Methods (by class)}{
\itemize{
\item \code{topoplot(default)}: Default method for data frames.

\item \code{topoplot(data.frame)}: Topographical plotting of data.frames and other non
\code{eeg_data} objects.

\item \code{topoplot(eeg_data)}: Topographical plotting of \code{eeg_data} objects.

\item \code{topoplot(eeg_epochs)}: Topographical plotting of \code{eeg_epochs} objects.

\item \code{topoplot(eeg_ICA)}: Topographical plot for \code{eeg_ICA} objects

\item \code{topoplot(eeg_tfr)}: Topographical plotting of \code{eeg_tfr} objects.

}}
\section{Notes on usage of Generalized Additive Models for interpolation}{
 The
function fits a GAM using the \code{gam} function from \code{mgcv}. Specifically, it
fits a spline using the model function \verb{gam(z ~ s(x, y, bs = "ts", k = 40)}. Using GAMs for smooths is very much experimental. The surface is
produced from the predictions of the GAM model fitted to the supplied data.
Values at each electrode do not necessarily match actual values in the
data: high-frequency variation will tend to be smoothed out. Thus, the
method should be used with caution. In addition the method can only be used
when there are more than 40 electrodes.
}

\examples{
topoplot(demo_epochs)
topoplot(demo_epochs, time_lim = c(.1, .2))
topoplot(demo_spatial, time_lim = list(0, .1, .2))
}
\seealso{
Other scalp-based maps: 
\code{\link{erp_scalp}()},
\code{\link{interactive_scalp}()}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
\concept{scalp-based maps}
