% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpad}
\alias{unpad}
\title{Unpad a vector}
\usage{
unpad(x, n)
}
\arguments{
\item{x}{Vector to unpad/trim}

\item{n}{Number of points to remove}
}
\description{
remove a specified number of points from the beginning and end of a vector
}
\keyword{internal}
