% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{validate_channels}
\alias{validate_channels}
\title{Chan_info checker}
\usage{
validate_channels(.data)
}
\arguments{
\item{.data}{an eeg_* object}
}
\description{
Checks for any missing channels in the chan_info; populates them with NA if
it finds any
}
\keyword{internal}
