% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{validate_eeg_data}
\alias{validate_eeg_data}
\title{Check eeg_data structure}
\usage{
validate_eeg_data(.data)
}
\description{
Check for missing fields and add them if necessary.
}
\keyword{internal}
