% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R
\name{validate_input}
\alias{validate_input}
\title{Validate input data for FASTER}
\usage{
validate_input(data)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}
}
\value{
The validated data object
}
\description{
This function checks if the input data is valid and has the required channel information.
}
\keyword{internal}
