% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_ica.R
\name{view_ica}
\alias{view_ica}
\title{EEG component viewer}
\usage{
view_ica(data)
}
\arguments{
\item{data}{An \code{eeg_ICA} object}
}
\description{
A Shiny viewer for ICA or SSD/RESS components that provides an interface for
looking at topographies, timecourses, and power spectral densities of all or
individual components.
}
\author{
Matt craddock \email{matt@mattcraddock.com}
}
