% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{wavelet_norm}
\alias{wavelet_norm}
\title{Normalise Morlet wavelets}
\usage{
wavelet_norm(mf_zp, n_freq)
}
\arguments{
\item{mf_zp}{A zero-padded, FFT'd morlet wavelet family}

\item{n_freq}{Number of frequencies}
}
\description{
Normalise the FFT'd Morlet wavelet family as recommended by Mike X Cohen,
dividing each wavelet by its absolute maximum. This should result in each
frequency being passed at unit amplitude, and the resulting convolution with
the signal should return units approximately on the original scale (i.e. uV^2
/ Hz)
}
\keyword{internal}
