% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{welch_fft}
\alias{welch_fft}
\title{Welch fft}
\usage{
welch_fft(data, seg_length, n_fft, noverlap, n_sig, srate)
}
\arguments{
\item{data}{Object to perform FFT on}

\item{seg_length}{length of each segment of data}

\item{n_fft}{length of FFT}

\item{noverlap}{overlap between segments}

\item{n_sig}{number of samples total}

\item{srate}{Sampling rate of the data}
}
\description{
Internal function for calculating the PSD using Welch's method
}
\keyword{internal}
