% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zero_vec}
\alias{zero_vec}
\title{Generate a zero centred vector}
\usage{
zero_vec(vec_length)
}
\arguments{
\item{vec_length}{how many points you want the vector to have}
}
\description{
Generate a zero centred vector
}
\keyword{internal}
