% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_for_ica.R
\name{ar_eogcor}
\alias{ar_eogcor}
\alias{ar_eogcor.eeg_ICA}
\title{Detect high component correlation with eye channels}
\usage{
ar_eogcor(decomp, data, ...)

\method{ar_eogcor}{eeg_ICA}(
  decomp,
  data,
  HEOG,
  VEOG,
  threshold = NULL,
  plot = TRUE,
  bipolarize = TRUE,
  ...
)
}
\arguments{
\item{decomp}{ICA decomposition}

\item{data}{Original data}

\item{...}{Other parameters}

\item{HEOG}{Horizontal eye channels}

\item{VEOG}{Vertical eye channels}

\item{threshold}{Threshold for correlation (r). Defaults to NULL,
automatically determining a threshold.}

\item{plot}{Plot correlation coefficient for all components}

\item{bipolarize}{Bipolarize the HEOG and VEOG channels?}
}
\value{
A character vector of component names that break the threshold.
}
\description{
Checks the correlation between individual components of an \code{eeg_ICA}
decomposition and the electrooculogram channels of an \code{eeg_epochs} dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_ICA}: Method for eeg_ICA objects.
}}

\references{
Chaumon, M., Bishop, D.V., Busch, N.A. (2015). A practical guide
to the selection of independent components of the electroencephalogram for
artifact correction. J Neurosci Methods. Jul 30;250:47-63. doi:
10.1016/j.jneumeth.2015.02.025
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
