% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{ar_eogcor}
\alias{ar_eogcor}
\alias{ar_eogcor.eeg_ICA}
\title{Detect high correlation with EOG channels}
\usage{
ar_eogcor(decomp, ...)

\method{ar_eogcor}{eeg_ICA}(decomp, data, HEOG, VEOG, threshold = NULL, plot = TRUE, ...)
}
\arguments{
\item{decomp}{ICA decomposition}

\item{...}{Other parameters}

\item{data}{Original data}

\item{HEOG}{Horizontal eye channels}

\item{VEOG}{Vertical eye channels}

\item{threshold}{Threshold for correlation (r). Defaults to NULL,
automatically determining a threshold.}

\item{plot}{Plot correlation coefficient for all components}
}
\description{
Detect high correlation with EOG channels
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_ICA}: Method for eeg_ICA objects.
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
