% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{ar_eogreg}
\alias{ar_eogreg}
\alias{ar_eogreg.eeg_data}
\alias{ar_eogreg.eeg_epochs}
\title{Remove EOG using regression}
\usage{
ar_eogreg(data, heog, veog, bipolarize = TRUE)

\method{ar_eogreg}{eeg_data}(data, heog, veog, bipolarize = TRUE)

\method{ar_eogreg}{eeg_epochs}(data, heog, veog, bipolarize = TRUE)
}
\arguments{
\item{data}{data to regress - \code{eeg_data} or \code{eeg_epochs}}

\item{heog}{Horizontal EOG channel labels}

\item{veog}{Vertical EOG channel labels}

\item{bipolarize}{Bipolarize the EOG channels. Only works when four channels
are supplied (2 HEOG and 2 VEOG).}
}
\description{
Calculates and removes the contribution of eye movements to the EEG signal using
least-squares regression.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
