% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_images.R
\name{arrange_chans}
\alias{arrange_chans}
\title{Topographical channel ordering}
\usage{
arrange_chans(chan_info, sig_names)
}
\arguments{
\item{chan_info}{channel info structure}

\item{sig_names}{names of signals in the data}
}
\value{
Vector of channel positions
}
\description{
Rearrange channels in a more anatomical order (e.g. left to right, anterior
to posterior).
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
\keyword{internal}
