% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_data}
\alias{as.data.frame.eeg_data}
\title{Convert eeg_data to data.frame}
\usage{
\method{as.data.frame}{eeg_data}(data, ..., long = FALSE)
}
\arguments{
\item{data}{Object of class \code{eeg_data}}

\item{...}{arguments for other as.data.frame commands}

\item{long}{Convert to long format. Defaults to FALSE}
}
\description{
Convert an object of class \code{eeg_data} into a standard data.frame / tibble
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
