% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_data}
\alias{as.data.frame.eeg_data}
\title{Convert eeg_data to data.frame}
\usage{
\method{as.data.frame}{eeg_data}(data, row.names = NULL, optional = FALSE,
  long = FALSE, ...)
}
\arguments{
\item{data}{Object of class \code{eeg_data}}

\item{row.names}{Kept for compatability with S3 method, ignored.}

\item{optional}{Kept for compatability with S3 method, ignored.}

\item{long}{Convert to long format. Defaults to FALSE}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert an object of class \code{eeg_data} into a standard data.frame / tibble
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
