% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.eeg_tfr}
\alias{as.data.frame.eeg_tfr}
\title{Convert \code{eeg_tfr} objects to data frames}
\usage{
\method{as.data.frame}{eeg_tfr}(x, row.names = NULL, optional = FALSE, long = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{eeg_tfr}}

\item{row.names}{Kept for compatability with S3 generic, ignored.}

\item{optional}{Kept for compatability with S3 generic, ignored.}

\item{long}{Convert to long format. Defaults to FALSE.}

\item{...}{arguments for other as.data.frame commands}
}
\description{
Convert \code{eeg_tfr} objects to data frames
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
