% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_averaging.R
\name{average_tf}
\alias{average_tf}
\title{Internal function for averaging over epochs for eeg_tfr objects.}
\usage{
average_tf(data)
}
\arguments{
\item{data}{data to average over}
}
\description{
Internal function for averaging over epochs for eeg_tfr objects.
}
\keyword{internal}
