% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_ica_ar.R
\name{bad_chans}
\alias{bad_chans}
\title{Identify globally bad channels}
\usage{
bad_chans(data, sds = 3, exclude = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_data}}

\item{sds}{Number of standard deviations of normalized data to consider as
acceptable range. Defaults to \code{3}.}

\item{exclude}{Channels to be ignored}

\item{...}{Further parameters (tbd)}
}
\value{
A character vector with channel names
}
\description{
Identify bad channels that should be excluded from ICA decomposition.
Channel autocorrelation values, variance and Hurst exponents are used to
determine bad channels.
\emph{See Reference for more details}
}
\references{
Nolan, Whelan & Reilly (2010). FASTER: Fully Automated Statistical Thresholding for
EEG artifact Rejection. J Neurosci Methods.
}
\seealso{
\code{\link[=bad_epochs]{bad_epochs()}}

Other Pre-ICA artifact rejection: 
\code{\link{bad_epochs}()}
}
\author{
Akash Mer
}
\concept{Pre-ICA artifact rejection}
