% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{channel_stats}
\alias{channel_stats}
\alias{channel_stats.eeg_data}
\title{Channel statistics}
\usage{
channel_stats(data, ...)

\method{channel_stats}{eeg_data}(data, ...)
}
\arguments{
\item{data}{Data as a \code{eeg_data} or \code{eeg_epochs} object.}

\item{...}{Other parameters passed to the functions.}
}
\value{
A data frame with statistics for each channel.
}
\description{
Channel statistics
}
\section{Methods (by class)}{
\itemize{
\item \code{channel_stats(eeg_data)}: Calculate channel statistics for \code{eeg_data}
objects.

}}
\examples{
channel_stats(demo_epochs)
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
