% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{check_timings}
\alias{check_timings}
\title{Check consistency of event and timing tables}
\usage{
check_timings(data)
}
\arguments{
\item{data}{\code{eeg_epochs} object}
}
\description{
Check consistency of event and timing tables
}
