% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{compute_g}
\alias{compute_g}
\title{Compute the g function for two sets of locations of channel locations on the
unit sphere.}
\usage{
compute_g(xyz_coords, xyz_elecs, m = 4, iter = 7)
}
\arguments{
\item{xyz_coords}{A set of electrode locations on a unit sphere.}

\item{xyz_elecs}{A set of electrode locations on a unit sphere.}

\item{m}{Interpolation constant (higher = less flexible)}
}
\description{
Compute the g function for two sets of locations of channel locations on the
unit sphere.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
