% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{compute_psd}
\alias{compute_psd}
\alias{compute_psd.eeg_data}
\title{Compute power spectral density}
\usage{
compute_psd(data, ...)

\method{compute_psd}{eeg_data}(data, seg_length = NULL, noverlap = 0,
  n_fft = 256, srate = NULL, method = "Welch")
}
\arguments{
\item{data}{Data to be plotted. Accepts objects of class \code{eeg_data}}

\item{...}{any further parameters passed to specific methods}

\item{seg_length}{Length of rolling data segments.}

\item{noverlap}{Number of (sampling) points of overlap between segments.}

\item{n_fft}{Length of FFT to be calculated.}

\item{srate}{Sampling rate}

\item{method}{Defaults to "Welch". No other method currently implemented.}
}
\description{
Returns PSD calculated using Welch's method for every channel in the data.
Currently concatenates across epochs. Returns uV^2 / Hz
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Compute PSD for an \code{eeg_data} object
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
