% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_analysis.R
\name{conv_mor}
\alias{conv_mor}
\title{Convolve with morlets}
\usage{
conv_mor(morlet_fam, signal, n, wavtime, srate)
}
\arguments{
\item{morlet_fam}{family of morlet wavelets}

\item{signal}{signal to be convolved}

\item{n}{points for FFT}

\item{wavtime}{time points}

\item{srate}{Sampling rate of the signal}
}
\description{
Convolve with morlets
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
