% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{conv_mor}
\alias{conv_mor}
\title{Convolve with morlets}
\usage{
conv_mor(morlet_fam, signal, n)
}
\arguments{
\item{morlet_fam}{family of morlet wavelets}

\item{signal}{signal to be convolved}

\item{n}{points for FFT}
}
\description{
Convolve with morlets
}
