% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{convert_tfr}
\alias{convert_tfr}
\title{Convert Fourier output to power or phase as requested.}
\usage{
convert_tfr(data, output)
}
\arguments{
\item{data}{Fourier coefficients from eeg_tfr}

\item{output}{What output is desired - "power", "phase" or "fourier"}
}
\description{
Convert Fourier output to power or phase as requested.
}
\keyword{internal}
