% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{create_tc}
\alias{create_tc}
\title{Internal function for creation of timecourse plots}
\usage{
create_tc(data, add_CI, colour, quantity = amplitude)
}
\arguments{
\item{data}{A data frame to be plotted}

\item{add_CI}{whether to add confidence intervals}

\item{colour}{whether to use colour}
}
\description{
Internal function for creation of timecourse plots
}
\keyword{internal}
