% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_averaging.R
\name{eeg_average}
\alias{eeg_average}
\alias{eeg_average.default}
\alias{eeg_average.eeg_epochs}
\alias{eeg_average.eeg_evoked}
\alias{eeg_average.eeg_tfr}
\title{Calculate averages (e.g. ERPs) for single datasets}
\usage{
eeg_average(data, ...)

\method{eeg_average}{default}(data, ...)

\method{eeg_average}{eeg_epochs}(data, cols = NULL, ...)

\method{eeg_average}{eeg_evoked}(data, cols = NULL, ...)

\method{eeg_average}{eeg_tfr}(data, cols = NULL, ...)
}
\arguments{
\item{data}{An \code{eeg_epochs} of \code{eeg_tfr} object.}

\item{...}{Other arguments passed to the averaging functions}

\item{cols}{Columns from the \code{epochs} structure that the average should
group on. NULL, the default, uses all columns other than the \code{epoch}
column.}
}
\value{
An object of class \code{eeg_evoked} if applied to \code{eeg_epochs};
\code{eeg_tfr} if applied to \code{eeg_tfr}.
}
\description{
This function is used to create an \code{eeg_evoked} object from \code{eeg_epochs}. By
default, it will try to keep different conditions in the data separate using
the \code{epochs} metadata from the object, thus yielding one average per
condition. Alternatively, the user can specify which averages they want using
the \code{cols} argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for averaging EEG objects

\item \code{eeg_epochs}: Create evoked data from \code{eeg_epochs}

\item \code{eeg_evoked}: average an \code{eeg_epochs} object over epochs.

\item \code{eeg_tfr}: average an \code{eeg_tfr} object over epochs.
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
