% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_averaging.R
\name{eeg_average}
\alias{eeg_average}
\alias{eeg_average.default}
\alias{eeg_average.eeg_epochs}
\alias{eeg_average.eeg_tfr}
\title{Calculate averages (e.g. ERPs) for single datasets}
\usage{
eeg_average(data, ...)

\method{eeg_average}{default}(data, ...)

\method{eeg_average}{eeg_epochs}(data, cols = NULL, ...)

\method{eeg_average}{eeg_tfr}(data, cols = NULL, ...)
}
\arguments{
\item{data}{An \code{eeg_epochs} object.}

\item{...}{Other arguments passed to the averaging functions}

\item{cols}{Columns from the \code{epochs} structure that the average should
group on. NULL, the default, uses all columns other than the \code{epoch}
column.}
}
\description{
This function is used to create an \code{eeg_evoked} object from
\code{eeg_epochs}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for averaging EEG objects

\item \code{eeg_epochs}: Create evoked data from \code{eeg_epochs}

\item \code{eeg_tfr}: average an eeg_tfr objects over epochs.
}}

\author{
Matt craddock \email{matt@mattcraddock.com}
}
