% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{eeg_average}
\alias{eeg_average}
\alias{eeg_average.eeg_epochs}
\title{Calculate averages (e.g. ERPs)}
\usage{
eeg_average(data, ...)

\method{eeg_average}{eeg_epochs}(data, cond_label = NULL, ...)
}
\arguments{
\item{data}{An \code{eeg_epochs} object.}

\item{...}{Other arguments passed to the averaging functions}

\item{cond_label}{Only pick events that include a given label. Character vector.}
}
\description{
Calculate averages (e.g. ERPs)

Create an \code{eeg_evoked} object from eeg_epochs
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Create evoked data from \code{eeg_epochs}
}}

\author{
Matt craddock \email{matt@mattcraddock.com}
}
