% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{eeg_combine}
\alias{eeg_combine}
\alias{eeg_combine.list}
\alias{eeg_combine.eeg_data}
\alias{eeg_combine.eeg_epochs}
\alias{eeg_combine.eeg_evoked}
\title{Combine \code{eegUtils} objects}
\usage{
eeg_combine(data, ...)

\method{eeg_combine}{list}(data, ...)

\method{eeg_combine}{eeg_data}(data, ..., check_timings = TRUE)

\method{eeg_combine}{eeg_epochs}(data, ..., check_timings = TRUE)

\method{eeg_combine}{eeg_evoked}(data, ...)
}
\arguments{
\item{data}{An \code{eeg_data}, \code{eeg_epochs}, or \code{eeg_evoked} object, or a list of
such objects.}

\item{...}{additional \code{eeg_data} or \code{eeg_epochs} objects}

\item{check_timings}{Check whether sample times / epoch numbers are
continuously ascending; if not, modify so that they are. Useful when, for
example, combining epochs derived from multiple recording blocks. Defaults to TRUE}
}
\value{
If all objects have the same \code{participant_id}, returns an object of
the same class as the original input object. If the objects have different
\code{participant_id} numbers, an object of both class \code{eeg_group} and the same
class as the original input object.
}
\description{
Combine multiple \code{eeg_epochs}, \code{eeg_data}, or \code{eeg_evoked} objects into a
single object. The function will check the \code{participant_id} entry in the
\code{epochs} structure of each object to see if the objects come from a single
participant or from multiple participants. If the data are from multiple
participants, it will create an \code{eeg_group} object. For individual
participants, it will check for duplicate epochs. For most objects, it will
concatenate the objects if none are found. However, for \code{eeg_data} it will
instead try to correct the epoch numbers. Check the details below for further
advice.
}
\details{
Combining \code{eeg_data} is mainly intended to be used for combining multiple
recordings from a single participant prior to subsequent epoching. Thus, the
function will change the epochs and timing structures of the resulting
combined object to be as if it were a single recording. The objects will be
combined in the input order, so ensure that the objects are input in
chronological order.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_combine(list)}: Method for combining lists of \code{eeg_data} and
\code{eeg_epochs} objects.

\item \code{eeg_combine(eeg_data)}: Method for combining \code{eeg_data} objects.

\item \code{eeg_combine(eeg_epochs)}: Method for combining \code{eeg_epochs} objects

\item \code{eeg_combine(eeg_evoked)}: Method for combining \code{eeg_evoked} objects

}}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
