% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eeg_data}
\alias{eeg_data}
\title{Function to create an S3 object of class "eeg_data".}
\usage{
eeg_data(data, srate, events = NULL, chan_info = NULL, timings = NULL,
  continuous = NULL, reference = NULL)
}
\arguments{
\item{data}{Raw data - signals from electrodes/channels.}

\item{srate}{Sampling rate in Hz.}

\item{chan_labels}{String of character names for electrodes.}
}
\description{
Function to create an S3 object of class "eeg_data".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
