% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eeg_data}
\alias{eeg_data}
\title{Function to create an S3 object of class "eeg_data".}
\usage{
eeg_data(data, srate, events = NULL, chan_labels = NULL, timings = NULL)
}
\arguments{
\item{data}{Raw data.}

\item{srate}{Sampling rate.}

\item{chan_labels}{}
}
\description{
Function to create an S3 object of class "eeg_data".
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
