% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{eeg_data}
\alias{eeg_data}
\title{Function to create an S3 object of class "eeg_data".}
\usage{
eeg_data(data, srate, events = NULL, chan_info = NULL,
  timings = NULL, continuous = NULL, reference = NULL)
}
\arguments{
\item{data}{Raw data - signals from electrodes/channels.}

\item{srate}{Sampling rate in Hz.}

\item{events}{Event table}

\item{chan_info}{String of character names for electrodes.}

\item{timings}{Timing information - samples and sample /samplirng rate.}

\item{continuous}{Whether the data is continuous or epoched.}

\item{reference}{Reference channel information, including names of reference
channels, excluded channels etc.}
}
\description{
Function to create an S3 object of class "eeg_data".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
