% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{eeg_filter}
\alias{eeg_filter}
\alias{eeg_filter.eeg_data}
\alias{eeg_filter.eeg_epochs}
\title{Filter EEG data}
\usage{
eeg_filter(.data, ...)

\method{eeg_filter}{eeg_data}(.data, low_freq = NULL, high_freq = NULL,
  filter_order = "auto", trans_bw = "auto", method = "fir",
  window = "hamming", ...)

\method{eeg_filter}{eeg_epochs}(.data, low_freq = NULL,
  high_freq = NULL, filter_order = "auto", trans_bw = "auto",
  method = "fir", window = "hamming", ...)
}
\arguments{
\item{.data}{An \code{eeg_data} or \code{eeg_epochs} object to be filtered}

\item{...}{Additional parameters.}

\item{low_freq}{Low cutoff frequency.}

\item{high_freq}{High cutoff frequency.}

\item{filter_order}{Defaults to "auto", which automatically estimates filter order for the specified filter characteristics (defaults to 4 if method = "iir").}

\item{trans_bw}{Transition bandwidth of the filter. "auto" or an integer.
"auto" attempts to determine a suitable transition bandwidth using the
heuristic given below. Ignored if method = "iir".}

\item{method}{"fir" (Finite Impulse Response) or "iir" (Infinite Impulse
Response). Defaults to "fir".}

\item{window}{Windowing function to use (FIR filtering only). Defaults to
"hamming"; currently only "hamming" available.}
}
\description{
Perform IIR or FIR filtering on input EEG data of class \code{eeg_data} or
\code{eeg_epochs}. WARNING: with epoched data, epoch boundaries are currently
ignored, which can result in minor edge artifacts.
}
\section{FIR versus IIR filtering}{
 Finite Impulse Response (FIR) filtering is
  performed using an overlap-add FFT method. Infinite Impulse Response (IIR)
  filtering is performed using a two-pass (once forwards, once reversed)
  method to correct for phase alignment.
}

