% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_handling.R
\name{eeg_ICA}
\alias{eeg_ICA}
\title{Function to create an S3 object of class "eeg_ICA".}
\usage{
eeg_ICA(mixing_matrix, unmixing_matrix, signals, timings, events,
  chan_info, srate, continuous)
}
\arguments{
\item{mixing_matrix}{ICA mixing matrix}

\item{unmixing_matrix}{ICA unmixing matrix}

\item{signals}{ICA components}

\item{timings}{Unique timepoints remaining in the data.}

\item{events}{event table}

\item{chan_info}{String of character names for electrodes.}

\item{srate}{Sampling rate}

\item{continuous}{Flag for whether data is continuous}
}
\description{
Function to create an S3 object of class "eeg_ICA".
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
