% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R, R/deprecated.R
\name{eeg_reference}
\alias{eeg_reference}
\alias{eeg_reference.default}
\alias{eeg_reference.eeg_data}
\alias{eeg_reference.eeg_epochs}
\alias{reref_eeg}
\title{Referencing}
\usage{
eeg_reference(data, ...)

\method{eeg_reference}{default}(data, ...)

\method{eeg_reference}{eeg_data}(
  data,
  ref_chans = "average",
  exclude = NULL,
  robust = FALSE,
  implicit_ref = NULL,
  verbose = TRUE,
  ...
)

\method{eeg_reference}{eeg_epochs}(
  data,
  ref_chans = "average",
  exclude = NULL,
  robust = FALSE,
  implicit_ref = NULL,
  verbose = TRUE,
  ...
)

reref_eeg(data, ...)
}
\arguments{
\item{data}{Data to re-reference. Primarily meant for use with data of class
\code{eeg_data}.}

\item{...}{Further parameters to be passed to \code{eeg_reference}}

\item{ref_chans}{Channels to reference data to. Defaults to "average" i.e.
average of all electrodes in data. Character vector of channel names or
numbers.}

\item{exclude}{Electrodes to exclude from average reference calculation.}

\item{robust}{Use median instead of mean; only used for average reference.
Defaults to FALSE.}

\item{implicit_ref}{Implicit reference channel - use this to add a channel
back that was previously used as a reference. E.g. if the LM (left mastoid)
channel was used in recording and is absent from the data, passing "LM"
adds an "LM" channel back to the data, populated with zeroes.}

\item{verbose}{Print informative messages in console. Defaults to TRUE.}
}
\value{
object of class \code{eeg_data}, re-referenced as requested.
}
\description{
Used to reference the EEG data to a specified electrode or electrodes.
Defaults to average reference. When specific electrodes are used, they are
removed from the data. Meta-data about the referencing scheme is held in the
\code{eeg_data} structure.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_reference(default)}: Default method

\item \code{eeg_reference(eeg_data)}: Rereference objects of class \code{eeg_data}

\item \code{eeg_reference(eeg_epochs)}: Rereference objects of class \code{eeg_epochs}

}}
\examples{
# demo_epochs is average referenced by default
demo_epochs
# Rereference it but exclude B5 from calculation of the average
eeg_reference(demo_epochs, exclude = "B5")
# Reference data using the median of the reference channels rather than the mean
eeg_reference(demo_epochs, robust = TRUE)

eeg_reference(demo_spatial)
eeg_reference(demo_spatial, ref_chans = "Fz")
eeg_reference(demo_spatial, implicit_ref = "LM")
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
