% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencing.R
\name{epoch_data}
\alias{epoch_data}
\title{Create epochs. Requires data of class \code{eeg_data}.}
\usage{
epoch_data(data, events, time_lim = (c(-1, 1)))
}
\arguments{
\item{data}{Continous data to be epoched.}

\item{events}{Character vector of events to epoch around.}

\item{time_lim}{Time in seconds to form epoch around the events. Defaults to one second either side.}
}
\description{
Create epochs. Requires data of class \code{eeg_data}.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
