% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epoching.R
\name{epoch_data}
\alias{epoch_data}
\alias{epoch_data.eeg_data}
\title{Create epochs from EEG data}
\usage{
epoch_data(data, ...)

\method{epoch_data}{eeg_data}(data, events, time_lim = c(-1, 1),
  baseline = NULL, epoch_labels = NULL, ...)
}
\arguments{
\item{data}{Continuous data to be epoched.}

\item{...}{Parameters passed to functions}

\item{events}{Character vector of events to epoch around.}

\item{time_lim}{Time in seconds to form epoch around the events. Defaults to
one second either side.}

\item{baseline}{Baseline times to subtract. Defaults to NULL (mean centres
epochs). Set to "none" to perform no corrections.}

\item{epoch_labels}{Character vector of same length as events which'll be
used to label the epochs.}
}
\value{
Returns an epoched object of class \code{eeg_epochs}
}
\description{
Creates epochs around specified event triggers. Requires data of class
\code{eeg_data}. Where multiple events are specified, epochs will be created
around each event.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Epoch \code{eeg_data} objects
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
