% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_images.R
\name{erp_image}
\alias{erp_image}
\title{Plot ERP images}
\usage{
erp_image(data, electrode = "Cz", smoothing = 10, clim = NULL)
}
\arguments{
\item{data}{Data frame to be plotted. Requires an amplitude column.}

\item{electrode}{electrode at which to generate an ERP image.}

\item{smoothing}{Number of trials to smooth over when generating image}

\item{clim}{Character vector of min and max values of plotting colour range. e.g. c(-5,5). Defaults to min and max.}
}
\description{
Plot ERP images
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
