% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoch_images.R
\name{erp_image}
\alias{erp_image}
\alias{erp_image.data.frame}
\alias{erp_image.eeg_epochs}
\alias{erp_image.eeg_ICA}
\alias{erp_image.eeg_tfr}
\title{Plot ERP images}
\usage{
erp_image(data, ...)

\method{erp_image}{data.frame}(
  data,
  electrode = "Cz",
  time_lim = NULL,
  smoothing = 10,
  clim = NULL,
  interpolate = FALSE,
  ...
)

\method{erp_image}{eeg_epochs}(
  data,
  electrode = "Cz",
  time_lim = NULL,
  smoothing = 10,
  clim = NULL,
  interpolate = FALSE,
  ...
)

\method{erp_image}{eeg_ICA}(
  data,
  component = "Comp001",
  smoothing = 10,
  clim = NULL,
  interpolate = FALSE,
  ...
)

\method{erp_image}{eeg_tfr}(
  data,
  electrode = "Cz",
  time_lim = NULL,
  smoothing = 10,
  clim = NULL,
  interpolate = FALSE,
  freq_range = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame to be plotted. Requires an amplitude column.}

\item{...}{Other arguments passed to the method.}

\item{electrode}{electrode at which to generate an ERP image.}

\item{time_lim}{Time limits of plot}

\item{smoothing}{Number of trials to smooth over when generating image}

\item{clim}{Character vector of min and max values of plotting colour range.
e.g. c(-5,5). Defaults to min and max.}

\item{interpolate}{Perform interpolation to produce smoother looking plots. Defaults to FALSE.}

\item{component}{\code{eeg_ICA} component to plot}

\item{freq_range}{A numeric vector specify the range of frequencies to
average over. A single number will find the closest matching frequency. A
vector of length two will match average over frequencies within that range.}
}
\value{
A \code{ggplot} object
}
\description{
Plot an ERP image from a single electrode. Smooths over a series of trials in
order to make across-trial patterns more apparent.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Default function operates on normal data frames

\item \code{eeg_epochs}: Create an \code{erp_image} from \code{eeg_epochs}

\item \code{eeg_ICA}: Plot component image from \code{eeg_ICA}

\item \code{eeg_tfr}: Plot component image from \code{eeg_tfr}
}}

\examples{
erp_image(demo_epochs, electrode = "A31")
erp_image(demo_epochs, electrode = "A31", interpolate = TRUE)
erp_image(demo_epochs, electrode = "A31", smoothing = 5)
erp_image(compute_tfr(demo_epochs,
 foi = c(4, 30), n_cycles = 3, n_freq = 20, verbose = FALSE, keep_trials = TRUE),
 electrode = "A31", freq_range = c(8, 12))
}
