% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_scalp.R
\name{erp_scalp}
\alias{erp_scalp}
\title{Plot ERPs on the scalp}
\usage{
erp_scalp(data, electrode = "electrode", amplitude = "amplitude",
  time = "time", color = NULL, size = 0.8, baseline = NULL,
  show_guide = TRUE, montage = NULL)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{Column name containing electrode names in data.
Defaults to "electrode".}

\item{amplitude}{Column name containing amplitudes in data.
Defaults to "amplitude".}

\item{time}{Column name containing time in data. Defaults to "time".}

\item{color}{Variable to color lines by. If no variable is passed, only
one line is drawn for each electrode.}

\item{size}{Size of line(s).}

\item{baseline}{Character vector of times to subtract for baseline correct.}

\item{show_guide}{Should a guide be shown.}

\item{montage}{Name of an existing montage set. Defaults to NULL; (currently
only 'biosemi64alpha' available other than default 10/20 system)}
}
\value{
Returns a ggplot2 plot object.
}
\description{
Creates an ERP figure for each electrode and layouts them on the scalp.
}
\details{
The function uses default electrode names and locations contained
in the package.
}
\examples{
erp_scalp(demo_epochs, montage = "biosemi64alpha")

}
\author{
Matti Vuorre, \email{mv2521@columbia.edu}
}
