% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_rejection.R
\name{faster_epo_stat}
\alias{faster_epo_stat}
\title{Calculate statistics for each channel in an epoch and identify bad channels}
\usage{
faster_epo_stat(data, exclude = NULL)
}
\arguments{
\item{data}{a matrix of signals from a single epoch}
}
\description{
Calculate statistics for each channel in an epoch and identify bad channels
}
\keyword{internal}
