% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{fft_n}
\alias{fft_n}
\title{N-point FFT}
\usage{
fft_n(signal, n)
}
\arguments{
\item{signal}{signal to be FFT'd}

\item{n}{Number of FFT points}
}
\description{
Either zero-pads or truncates a signal to N and runs an FFT
}
