% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_coefs}
\alias{filter_coefs}
\title{Generate filter coefficients}
\usage{
filter_coefs(method, filt_pars, filter_order, window)
}
\arguments{
\item{method}{IIR or FIR}

\item{filt_pars}{output of parse_filt_freqs}

\item{filter_order}{order of the filter in samples}

\item{window}{Ignored for IIR}
}
\description{
Generate filter coefficients for an IIR or FIR filter.
}
\keyword{internal}
