% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\name{geom_topo}
\alias{geom_topo}
\title{Create a topographical plot}
\usage{
geom_topo(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = TRUE,
  inherit.aes = TRUE,
  interpolate = FALSE,
  interp_limit = "skirt",
  chan_markers = "point",
  chan_size = rel(2),
  head_size = rel(1.5),
  r = NULL,
  grid_res = 200,
  method = "biharmonic",
  bins = 6,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{interp_limit}{Topoplot with a "skirt" or inside the "head".}

\item{chan_markers}{Defaults to "point". Mark electrode positions with points
or text.}

\item{chan_size}{Size for channel markers, if any.}

\item{head_size}{Size of the head shape.}

\item{r}{Head circumference}

\item{grid_res}{Smoothness of the interpolation grid.}

\item{method}{"biharmonic" or ""gam".}

\item{bins}{Number of bins to use for contour lines.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\code{geom_topo()} creates a topographical plot as a \code{ggplot2} object.
This function automatically combines a number of distinct geom_* and stat_*
functions to create a default topographical scalp map. Since geom_raster does
not allow unevenly spaced grids, the function creates an interpolated surface.
}
\examples{
library(ggplot2)
ggplot(demo_epochs, aes(x = x, y = y, fill = amplitude, z = amplitude)) + geom_topo()
}
\seealso{
Other topoplot functions: 
\code{\link{stat_scalpcontours}()},
\code{\link{stat_scalpmap}()}
}
\concept{topoplot functions}
