% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fitting.R
\name{get_lm}
\alias{get_lm}
\title{Calculate coefficients and standard errors}
\usage{
get_lm(mdf, x, inverted_dm, resid_df)
}
\arguments{
\item{mdf}{model matrix}

\item{x}{data to be modelled (matrix of trials x channels)}

\item{inverted_dm}{inverted matrix}
}
\description{
Calculate coefficients and standard errors
}
\keyword{internal}
