% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\name{get_scalpmap}
\alias{get_scalpmap}
\alias{get_scalpmap.data.frame}
\title{Get an interpolated grid}
\usage{
get_scalpmap(data, ...)

\method{get_scalpmap}{data.frame}(
  data,
  method = "biharmonic",
  grid_res = 100,
  interp_limit = "skirt",
  quantity = "amplitude",
  facets = NULL,
  ...
)
}
\arguments{
\item{data}{An object of class 'data.frame`}

\item{...}{Other arguments}

\item{method}{"biharmonic" or "gam"}

\item{grid_res}{grid resolution}

\item{interp_limit}{interpolate up to the "head" or add a "skirt"}

\item{quantity}{"amplitude"}

\item{facets}{Any facets you plan to use}
}
\description{
Get an interpolated grid
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: data.frame
}}

