% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scalpmap.R
\name{get_scalpmap}
\alias{get_scalpmap}
\alias{get_scalpmap.data.frame}
\title{Get an interpolated grid}
\usage{
get_scalpmap(data, ...)

\method{get_scalpmap}{data.frame}(
  data,
  method = "biharmonic",
  grid_res = 100,
  interp_limit = "skirt",
  quantity = "amplitude",
  facets = NULL,
  r = 95,
  ...
)
}
\arguments{
\item{data}{An object of class 'data.frame`}

\item{...}{Other arguments}

\item{method}{"biharmonic" or "gam" smooth}

\item{grid_res}{Grid resolution}

\item{interp_limit}{Interpolate up to the "head" or add a "skirt"}

\item{quantity}{Quantity to be plotted. Defaults to "amplitude".}

\item{facets}{Any facets you plan to use}

\item{r}{Size of headshape}
}
\description{
Get an interpolated grid
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: data.frame
}}

