% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scalpmap.R
\name{get_scalpmap}
\alias{get_scalpmap}
\alias{get_scalpmap.data.frame}
\alias{get_scalpmap.eeg_epochs}
\title{Calculate an interpolated scalpmap}
\usage{
get_scalpmap(data, ...)

\method{get_scalpmap}{data.frame}(
  data,
  method = "biharmonic",
  grid_res = 100,
  interp_limit = "skirt",
  quantity = "amplitude",
  facets = NULL,
  r = 95,
  k = -1,
  ...
)

\method{get_scalpmap}{eeg_epochs}(
  data,
  method = "biharmonic",
  grid_res = 100,
  interp_limit = "skirt",
  quantity = "amplitude",
  facets = NULL,
  r = 95,
  ...
)
}
\arguments{
\item{data}{An object of class 'data.frame`}

\item{...}{Other arguments}

\item{method}{"biharmonic" or "gam" smooth}

\item{grid_res}{Grid resolution}

\item{interp_limit}{Interpolate up to the "head" or add a "skirt"}

\item{quantity}{Quantity to be plotted. Defaults to "amplitude".}

\item{facets}{Any facets you plan to use}

\item{r}{Size of headshape}

\item{k}{Degrees of freedom used for spline when using \code{method = gam}.
Defaults to -1, which attempts to automatically determine k.}
}
\value{
A \code{tibble} with the columns \code{x}, \code{y}, \code{fill}
}
\description{
Calculate and return an interpolated scalpmap from a \code{data.frame} or
\code{eeg_epochs} object. This provides the basis for a raster plot, as used to
create topographical plots.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_scalpmap(data.frame)}: Get scalpmap from a \code{data.frame}

\item \code{get_scalpmap(eeg_epochs)}: Get scalpmap from an \code{eeg_epochs} object.

}}
\examples{
head(get_scalpmap(demo_epochs))
}
