% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{iir_filt}
\alias{iir_filt}
\title{Butterworth IIR filter}
\usage{
iir_filt(data, low_freq = NULL, high_freq = NULL, filter_order = 4,
  srate = NULL, plot_filt = FALSE, silent = FALSE)
}
\arguments{
\item{data}{Data to be filtered.}

\item{low_freq}{Low passband edge.}

\item{high_freq}{High passband edge.}

\item{filter_order}{Order of the Butterworth filter.}

\item{srate}{Sampling rate of the signal.}

\item{plot_filt}{Plot filter characteristics}

\item{silent}{Turns off filtering messages.}
}
\description{
Construct a Butterworth IIR filter and filter input data. This uses \code{signal::filt_filt}, which filters the signal twice to - once forwards, then again backwards)
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
