% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{iir_filt}
\alias{iir_filt}
\alias{iir_filt.data.frame}
\alias{iir_filt.eeg_data}
\alias{iir_filt.eeg_epochs}
\title{Butterworth IIR filter}
\usage{
iir_filt(data, ...)

\method{iir_filt}{data.frame}(data, low_freq = NULL, high_freq = NULL,
  filter_order = 4, srate, silent = FALSE, ...)

\method{iir_filt}{eeg_data}(data, low_freq = NULL, high_freq = NULL,
  filter_order = 4, silent = FALSE, ...)

\method{iir_filt}{eeg_epochs}(data, low_freq = NULL, high_freq = NULL,
  filter_order = 4, silent = FALSE, ...)
}
\arguments{
\item{data}{Data to be filtered.}

\item{...}{Parameters passed to S3 methods}

\item{low_freq}{Low passband edge.}

\item{high_freq}{High passband edge.}

\item{filter_order}{Order of the Butterworth filter.}

\item{srate}{Sampling rate of the signal.}

\item{silent}{Turns off filtering messages.}
}
\description{
Construct a Butterworth IIR filter and filter input data. This uses
\code{signal::filt_filt}, which filters the signal twice to - once forwards,
then again backwards).
}
\details{
low_freq and high_freq are passband edges. Pass low freq or high freq alone
to perform high-pass or low-pass filtering respectively. For band-pass or
band-stop filters, pass both low_freq and high_freq.

If low_freq < high_freq, bandpass filtering is performed.

If low_freq > high_freq, bandstop filtering is performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Filter a data frame

\item \code{eeg_data}: Filter eeg_data

\item \code{eeg_epochs}: Filter eeg_epochs
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
