% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{import_chans}
\alias{import_chans}
\title{Import channel locations from various file formats}
\usage{
import_chans(file_name, format = "spherical")
}
\arguments{
\item{file_name}{Name and full path of file to be loaded.}

\item{format}{If the file is not .elc format, "spherical", "geographic".
Default is "spherical".}
}
\description{
Currently only ASA .elc format with Cartesian x-y-z coordinates is supported.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
