% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_ft}
\alias{import_ft}
\title{Import Fieldtrip files}
\usage{
import_ft(file_name, participant_id = NULL, recording = NULL, verbose = TRUE)
}
\arguments{
\item{file_name}{Name of file to be imported.}

\item{participant_id}{Identifier for the participant.}

\item{recording}{Name of the recording. By default, the filename will be
used.}

\item{verbose}{Informative messages printed to console. Defaults to TRUE.}
}
\description{
Fieldtrip is a Matlab package for EEG/MEG processing and analysis.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
