% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_raw}
\alias{import_raw}
\title{Function for reading raw data.}
\usage{
import_raw(file_name, file_path = NULL, chan_nos = NULL)
}
\arguments{
\item{file_name}{File to import.}

\item{file_path}{Path to file name, if not included in filename.}

\item{chan_nos}{Channels to import. All channels are included by default.}
}
\description{
Currently only BDF or EDF files are supported. The \code{edfReader} package reads the data in, and then the function creates an eeg_data structure for subsequent use.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
