% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{import_raw}
\alias{import_raw}
\title{Function for reading raw data.}
\usage{
import_raw(
  file_name,
  file_path = NULL,
  chan_nos = NULL,
  recording = NULL,
  participant_id = character(1),
  fast_bdf = TRUE
)
}
\arguments{
\item{file_name}{File to import. Should include file extension.}

\item{file_path}{Path to file name, if not included in filename.}

\item{chan_nos}{Channels to import. All channels are included by default.}

\item{recording}{Name of the recording. By default, the filename will be
used.}

\item{participant_id}{Identifier for the participant.}

\item{fast_bdf}{New, faster method for loading BDF files. Experimental.}
}
\description{
Currently BDF/EDF, 32-bit .CNT, and Brain Vision Analyzer files are
supported. Filetype is determined by the file extension.The \code{edfReader}
package is used to load BDF/EDF files, whereas custom code is used for .CNT
and BVA files. The function creates an \code{eeg_data} structure for
subsequent use.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
