% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erp_scalp.R
\name{interactive_scalp}
\alias{interactive_scalp}
\title{Interactive scalp maps}
\usage{
interactive_scalp(data, colour = NULL)
}
\arguments{
\item{data}{An EEG dataset.}

\item{colour}{Variable to colour lines by. If no variable is passed, only
one line is drawn for each electrode.}
}
\description{
Launches a Shiny Gadget for an interactive version of erp_scalp, allowing clicking of individual electrodes to plot them in a separate panel. In that panel, they can be averaged over or plotted as individual electrodes.
Currently buggy with datasets with existing electrode co-ordinates.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
