% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_scalp.R
\name{interactive_scalp}
\alias{interactive_scalp}
\title{Interactive scalp maps}
\usage{
interactive_scalp(data, colour = NULL)
}
\arguments{
\item{data}{An EEG dataset.}
}
\description{
Launchs a Shiny Gadget for an interactive version of erp_scalp, allowing clicking of individual electrodes to plot them separately.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
