% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{interp_elecs}
\alias{interp_elecs}
\alias{interp_elecs.eeg_data}
\title{Channel interpolation}
\usage{
interp_elecs(data, bad_elecs, ...)

\method{interp_elecs}{eeg_data}(data, bad_elecs, ...)
}
\arguments{
\item{data}{Data as a \code{eeg_data} or \code{eeg_epochs} object.}

\item{bad_elecs}{Name(s) of electrode(s) to interpolate.}

\item{...}{Other parameters passed to the functions.}
}
\description{
Interpolate EEG channels using a spherical spline (Perrin et al., 1989; 1990). The
data must have channel locations attached.
}
\section{Methods (by class)}{
\itemize{
\item \code{interp_elecs(eeg_data)}: Interpolate EEG channel(s)

}}
\references{
\itemize{
\item Perrin, F., Pernier, J., Bertrand, O., & Echallier, J. F.
(1989). Spherical splines for scalp potential and current
density mapping. Electroencephalography and Clinical
Neurophysiology, 72, 184-187
\item Perrin, F., Pernier, J., Bertrand, O., & Echallier, J. F.
(1990). Corrigenda EEG 02274. Electroencephalography and
Clinical Neurophysiology, 76, 565
}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
