% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_codes.R
\name{list_epochs}
\alias{list_epochs}
\alias{list_epochs.eeg_epochs}
\alias{list_epochs.eeg_ICA}
\title{List epochs}
\usage{
list_epochs(data, ...)

\method{list_epochs}{eeg_epochs}(data, ...)

\method{list_epochs}{eeg_ICA}(data, ...)
}
\arguments{
\item{data}{An object of class \code{eeg_epochs}}

\item{...}{Additional arguments}
}
\description{
List trigger types and any labels found in an \code{eeg_epochs} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{list_epochs(eeg_epochs)}: List epochs and associated events from \code{eeg_epochs} objects

\item \code{list_epochs(eeg_ICA)}: List epochs and associated events from \code{eeg_ICA} objects

}}
\seealso{
\code{\link[=tag_events]{tag_events()}} and \code{\link[=list_events]{list_events()}}

Other event handlers: 
\code{\link{events}()},
\code{\link{list_events}()},
\code{\link{tag_events}()}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\concept{event handlers}
