% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{morlet}
\alias{morlet}
\title{Morlet wavelet}
\usage{
morlet(frex, srate, n_cycles, n_freq)
}
\arguments{
\item{frex}{Frequency range of interest}

\item{n_cycles}{Length of wavelet in cycles}

\item{n_freq}{number of frequencies to resolve}
}
\description{
Generate Morlet wavelet family
}
\keyword{internal}
