% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{morlet}
\alias{morlet}
\title{Morlet wavelet}
\usage{
morlet(frex, wavtime, n_cycles = 7, n_freq = 30)
}
\arguments{
\item{frex}{Frequency range of interest}

\item{wavtime}{Times at which to place a wavelet}

\item{n_cycles}{Length of wavelet in cycles}

\item{n_freq}{number of frequencies to resolve}
}
\description{
Generate Morlet wavelet family
}
