% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad}
\alias{pad}
\title{Pad a vector with zeros}
\usage{
pad(x, n)
}
\arguments{
\item{x}{vector to pad}

\item{n}{number of zeros to pad}
}
\description{
Pads a vector with zeros at the beginning and end.
}
\keyword{internal}
