% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{parse_chaninfo}
\alias{parse_chaninfo}
\title{Parse channel info from an EEGLAB set file}
\usage{
parse_chaninfo(chan_info)
}
\arguments{
\item{chan_info}{Channel info list from an EEGLAB set file}
}
\description{
Internal function to convert EEGLAB chan_info to eegUtils style
}
\keyword{internal}
