% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{parse_for_tc}
\alias{parse_for_tc}
\title{Parse data for timecourses}
\usage{
parse_for_tc(data, time_lim, electrode, baseline, add_CI)
}
\arguments{
\item{data}{data to be parsed}

\item{time_lim}{time limits to be returned.}

\item{electrode}{electrodes to be selected}

\item{baseline}{baseline times to be average and subtracted}

\item{add_CI}{Logical for whether CIS are required}
}
\description{
Internal command for parsing various data structures into a suitable format
for \code{tc_plot}
}
\keyword{internal}
