% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_io.R
\name{parse_vhdr_chans}
\alias{parse_vhdr_chans}
\title{Parse BVA channel info}
\usage{
parse_vhdr_chans(chan_labels, chan_info)
}
\arguments{
\item{chan_labels}{The `Channel Infos` section from a BVA vhdr file}

\item{chan_info}{The `Coordinates` section from a BVA vhdr file}
}
\description{
Parse BVA channel info
}
\keyword{internal}
