% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_butterfly.R
\name{plot_butterfly}
\alias{plot_butterfly}
\alias{plot_butterfly.default}
\alias{plot_butterfly.eeg_evoked}
\alias{plot_butterfly.eeg_data}
\alias{plot_butterfly.eeg_epochs}
\alias{plot_butterfly.eeg_stats}
\alias{plot_butterfly.eeg_lm}
\title{Create a butterfly plot from timecourse data}
\usage{
plot_butterfly(data, ...)

\method{plot_butterfly}{default}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  continuous = FALSE,
  browse_mode = FALSE,
  allow_facets = FALSE,
  ...
)

\method{plot_butterfly}{eeg_evoked}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  continuous = FALSE,
  browse_mode = FALSE,
  allow_facets = FALSE,
  ...
)

\method{plot_butterfly}{eeg_data}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  allow_facets = FALSE,
  browse_mode = FALSE,
  ...
)

\method{plot_butterfly}{eeg_epochs}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  allow_facets = FALSE,
  browse_mode = FALSE,
  ...
)

\method{plot_butterfly}{eeg_stats}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  allow_facets = FALSE,
  browse_mode = FALSE,
  quantity = "statistic",
  ...
)

\method{plot_butterfly}{eeg_lm}(
  data,
  time_lim = NULL,
  baseline = NULL,
  legend = TRUE,
  allow_facets = FALSE,
  browse_mode = FALSE,
  quantity = "coefficients",
  ...
)
}
\arguments{
\item{data}{EEG dataset. Should have multiple timepoints.}

\item{...}{Other parameters passed to \code{plot_butterfly}}

\item{time_lim}{Character vector. Numbers in whatever time unit is used
specifying beginning and end of time-range to plot. e.g. c(-.1,.3)}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean
over the specified period and subtracts. e.g. c(-.1, 0)}

\item{legend}{Include plot legend. Defaults to TRUE.}

\item{continuous}{Is the data continuous or not (I.e. epoched)}

\item{browse_mode}{Applies custom theme used with \code{browse_data()}.}

\item{allow_facets}{Allow use of ggplot2 facetting. See note below. Defaults
to FALSE.}

\item{quantity}{Which aspect of the linear model you want to be plotted. only
applies to \code{eeg_lm} objects}
}
\value{
ggplot2 object showing ERPs for all electrodes overlaid on a single
plot.
}
\description{
Typically event-related potentials/fields, but could also be timecourses from
frequency analyses for single frequencies. Output is a ggplot2 object. CIs
not possible.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_butterfly(default)}: Default \code{plot_butterfly} method for data.frames,
\code{eeg_data}

\item \code{plot_butterfly(eeg_evoked)}: Plot butterfly for \code{eeg_evoked} objects

\item \code{plot_butterfly(eeg_data)}: Create butterfly plot for \code{eeg_data} objects

\item \code{plot_butterfly(eeg_epochs)}: Create butterfly plot for \code{eeg_epochs} objects

\item \code{plot_butterfly(eeg_stats)}: Create butterfly plot for \code{eeg_stats} objects

\item \code{plot_butterfly(eeg_lm)}: Create butterfly plot for \code{eeg_lm} objects

}}
\section{Notes on ggplot2 facetting}{


In order for ggplot2 facetting to work, the data has to be plotted using
\code{stat_summary()} rather than \code{geom_line()}, so that the plots can still be
made when not all categorical variables are reflected in the facets. e.g.
if there are two variables with two levels each, but you want to average
over one of those variables, \code{stat_summary()} is required. However,
\code{stat_summary()} can be extremely slow with a large number of timepoints,
electrodes, and epochs.
}

\examples{
plot_butterfly(demo_epochs)
plot_butterfly(demo_epochs,
time_lim = c(-.1, .4),
legend = FALSE)
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
