% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{plot_butterfly}
\alias{plot_butterfly}
\alias{plot_butterfly.default}
\alias{plot_butterfly.eeg_evoked}
\alias{plot_butterfly.eeg_stats}
\alias{plot_butterfly.eeg_data}
\alias{plot_butterfly.eeg_epochs}
\title{Create a butterfly plot from timecourse data}
\usage{
plot_butterfly(data, ...)

\method{plot_butterfly}{default}(data, time_lim = NULL,
  baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, facet, ...)

\method{plot_butterfly}{eeg_evoked}(data, time_lim = NULL,
  baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, facet, ...)

\method{plot_butterfly}{eeg_stats}(data, time_lim = NULL,
  baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, ...)

\method{plot_butterfly}{eeg_data}(data, time_lim = NULL,
  baseline = NULL, legend = TRUE, facet, browse_mode = FALSE, ...)

\method{plot_butterfly}{eeg_epochs}(data, time_lim = NULL,
  baseline = NULL, legend = TRUE, facet, browse_mode = FALSE, ...)
}
\arguments{
\item{data}{EEG dataset. Should have multiple timepoints.}

\item{...}{Other parameters passed to plot_butterfly}

\item{time_lim}{Character vector. Numbers in whatever time unit is used
specifying beginning and end of time-range to plot. e.g. c(-.1,.3)}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean
over the specified period and subtracts. e.g. c(-.1, 0)}

\item{colourmap}{Attempt to plot using a different colourmap (from
RColorBrewer). (Not yet implemented)}

\item{legend}{Plot legend or not.}

\item{continuous}{Is the data continuous or not (I.e. epoched)}

\item{browse_mode}{Custom theme for use with browse_data.}

\item{facet}{Deprecated. Please use standard ggplot2 facetting functions.}
}
\value{
ggplot2 object showing ERPs for all electrodes overlaid on a single
  plot.
}
\description{
Typically event-related potentials/fields, but could also be timecourses from
frequency analyses for single frequencies. Output is a ggplot2 object. CIs
not possible.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default `plot_butterfly` method for data.frames, \code{eeg_data}

\item \code{eeg_evoked}: Plot butterfly for \code{eeg_evoked} objects

\item \code{eeg_stats}: Butterfly plot for EEG statistics

\item \code{eeg_data}: Create butterfly plot for \code{eeg_data} objects

\item \code{eeg_epochs}: Create butterfly plot for \code{eeg_epochs} objects
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
