% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{plot_electrodes}
\alias{plot_electrodes}
\alias{plot_electrodes.default}
\alias{plot_electrodes.eeg_data}
\title{Plot electrode locations}
\usage{
plot_electrodes(data, interact = FALSE)

\method{plot_electrodes}{default}(data, interact = FALSE)

\method{plot_electrodes}{eeg_data}(data, interact = FALSE)
}
\arguments{
\item{data}{Data with associated electrode locations to be plotted.}

\item{interact}{Choose 2D cartesian layout, or, if set to TRUE, an
interactive 3D plot of electrode locations. Defaults to FALSE.}
}
\description{
Produces either a 2D plot of the electrode locations or an interactive plot
of electrode locations in 3D space.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: generic plot electrodes function

\item \code{eeg_data}: Plot electrodes associated with an \code{eeg_data} object.
}}

\examples{

plot_electrodes(demo_epochs)

}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
