% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_plotting.R
\name{plot_psd}
\alias{plot_psd}
\alias{plot_psd.eeg_epochs}
\title{Plot power spectral density}
\usage{
plot_psd(data, freq_range = NULL, ...)

\method{plot_psd}{eeg_epochs}(data, freq_range = NULL, ...)
}
\arguments{
\item{data}{Object of class \code{eeg_epochs}}

\item{freq_range}{Vector of lower and upper frequencies to plot. (e.g. c(1, 40))}

\item{...}{Additional parameters}
}
\description{
Plot power spectral density
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Plot PSDs for \code{eeg_epochs}.
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
