% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{plot_timecourse}
\alias{plot_timecourse}
\alias{plot_timecourse.data.frame}
\alias{plot_timecourse.eeg_evoked}
\alias{plot_timecourse.eeg_ICA}
\alias{plot_timecourse.eeg_epochs}
\alias{plot_timecourse.eeg_group}
\alias{plot_timecourse.eeg_tfr}
\title{Plot one-dimensional timecourse data.}
\usage{
plot_timecourse(data, ...)

\method{plot_timecourse}{data.frame}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  facets = NULL,
  ...
)

\method{plot_timecourse}{eeg_evoked}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  facets = NULL,
  ...
)

\method{plot_timecourse}{eeg_ICA}(
  data,
  component = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  facets = NULL,
  ...
)

\method{plot_timecourse}{eeg_epochs}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  facets = NULL,
  ...
)

\method{plot_timecourse}{eeg_group}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  facets = NULL,
  ...
)

\method{plot_timecourse}{eeg_tfr}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  freq_range = NULL,
  type = "divide",
  ...
)
}
\arguments{
\item{data}{EEG dataset. Should have multiple timepoints.}

\item{...}{Other arguments passed to methods.}

\item{electrode}{Electrode(s) to plot.}

\item{time_lim}{Character vector. Numbers in whatever time unit is used
specifying beginning and end of time-range to plot. e.g. c(-.1, .3)}

\item{add_CI}{Add confidence intervals to the graph. Defaults to 95 percent
between-subject CIs.}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean
over the specified period and subtracts. e.g. c(-.1,0)}

\item{colour}{Variable to colour lines by. If no variable is passed, only one
line is drawn.}

\item{color}{Alias for colour.}

\item{mapping}{A \code{ggplot2} \code{aes()} mapping.}

\item{facets}{A right-hand-side only formula specifying which variables should
be used to create facets.}

\item{component}{name or number of ICA component to plot}

\item{freq_range}{Choose a specific frequency range to plot. If NULL,
calculates the mean over all frequencies. Note that this does not imply
that there is power at an included frequency. For example, lower
frequencies will have shorter timecourses than high frequencies.}

\item{type}{Type of baseline correction to use for \code{eeg_tfr} objects}
}
\value{
Returns a ggplot2 plot object
}
\description{
Typically event-related potentials/fields, but could also be timecourses from
frequency analyses for single frequencies. Averages over all submitted
electrodes. For group data, \code{plot_timecourse} will average within-participants
first, using weighted averaging where possible, then across participants using
unweighted averaging. Output is a \code{ggplot2} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_timecourse(data.frame)}: Plot a data.frame timecourse

\item \code{plot_timecourse(eeg_evoked)}: plot \code{eeg_evoked} timecourses

\item \code{plot_timecourse(eeg_ICA)}: Plot individual components from \code{eeg_ICA}
components

\item \code{plot_timecourse(eeg_epochs)}: Plot timecourses from \code{eeg_epochs} objects.

\item \code{plot_timecourse(eeg_group)}: Plot timecourses from \code{eeg_group} objects.

\item \code{plot_timecourse(eeg_tfr)}: Plot timecourses from \code{eeg_tfr} objects.

}}
\examples{
library(ggplot2)
plot_timecourse(demo_epochs, "A29")
plot_timecourse(demo_epochs, "A29", add_CI = TRUE)
plot_timecourse(demo_spatial, "Oz", mapping = aes(colour = epoch_labels))
plot_timecourse(demo_spatial, "Oz", facets = ~epoch_labels)
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
