% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timecourse.R
\name{plot_timecourse}
\alias{plot_timecourse}
\alias{plot_timecourse.data.frame}
\alias{plot_timecourse.eeg_evoked}
\alias{plot_timecourse.eeg_ICA}
\alias{plot_timecourse.eeg_epochs}
\alias{plot_timecourse.eeg_group}
\alias{plot_timecourse.eeg_tfr}
\title{Plot 1-D timecourse data.}
\usage{
plot_timecourse(data, ...)

\method{plot_timecourse}{data.frame}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  ...
)

\method{plot_timecourse}{eeg_evoked}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  ...
)

\method{plot_timecourse}{eeg_ICA}(
  data,
  component = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  ...
)

\method{plot_timecourse}{eeg_epochs}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  ...
)

\method{plot_timecourse}{eeg_group}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  ...
)

\method{plot_timecourse}{eeg_tfr}(
  data,
  electrode = NULL,
  time_lim = NULL,
  add_CI = FALSE,
  baseline = NULL,
  colour = NULL,
  color = NULL,
  mapping = NULL,
  freq_range = NULL,
  ...
)
}
\arguments{
\item{data}{EEG dataset. Should have multiple timepoints.}

\item{...}{Other arguments passed to methods.}

\item{electrode}{Electrode(s) to plot.}

\item{time_lim}{Character vector. Numbers in whatever time unit is used
specifying beginning and end of time-range to plot. e.g. c(-.1, .3)}

\item{add_CI}{Add confidence intervals to the graph. Defaults to 95 percent
between-subject CIs.}

\item{baseline}{Character vector. Times to use as a baseline. Takes the mean
over the specified period and subtracts. e.g. c(-.1,0)}

\item{colour}{Variable to colour lines by. If no variable is passed, only one
line is drawn.}

\item{color}{Alias for colour.}

\item{mapping}{A ggplot2 \code{aes()} mapping.}

\item{component}{name or number of ICA component to plot}

\item{freq_range}{Choose a specific frequency range to plot}
}
\value{
Returns a ggplot2 plot object
}
\description{
Typically event-related potentials/fields, but could also be timecourses from
frequency analyses for single frequencies. Averages over all submitted
electrodes. Output is a ggplot2 object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Plot a data.frame timecourse

\item \code{eeg_evoked}: plot \code{eeg_evoked} timecourses

\item \code{eeg_ICA}: Plot individual components from \code{eeg_ICA} components

\item \code{eeg_epochs}: Plot timecourses from \code{eeg_epochs} objects.

\item \code{eeg_group}: Plot timecourses from \code{eeg_group} objects.

\item \code{eeg_tfr}: Plot timecourses from \code{eeg_tfr} objects.
}}

\examples{
plot_timecourse(demo_epochs, "A29")
plot_timecourse(demo_epochs, "A29", add_CI = TRUE)
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
