% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar-faster.R, R/pre_ica_ar.R
\name{quick_hurst}
\alias{quick_hurst}
\title{Quickly calculate simple Hurst exponent for a matrix}
\usage{
quick_hurst(data)

quick_hurst(data)
}
\arguments{
\item{data}{matrix of EEG signals}
}
\description{
Quickly calculate simple Hurst exponent for a matrix

Quickly calculate simple Hurst exponent for a matrix
}
\keyword{internal}
