% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{rearrange_tfr}
\alias{rearrange_tfr}
\title{Rearrange and combine \code{tfr_average} objects}
\usage{
rearrange_tfr(data, add_obj)
}
\arguments{
\item{data}{first object to be combined}

\item{add_obj}{list of additional objects}
}
\description{
Epochs in different \code{tfr_average} objects are not necessarily in the same order.
}
\keyword{internal}
