% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-extensions.R, R/ggplot2-extensions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{select}
\alias{mutate}
\alias{rename}
\alias{fortify}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}}

  \item{ggplot2}{\code{\link[ggplot2]{fortify}}}
}}

