% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{reref_eeg}
\alias{reref_eeg}
\alias{reref_eeg.default}
\alias{reref_eeg.eeg_data}
\title{Referencing}
\usage{
reref_eeg(data, ...)

\method{reref_eeg}{default}(data, ...)

\method{reref_eeg}{eeg_data}(data, ref_chans = "average", exclude = NULL,
  robust = FALSE, ...)
}
\arguments{
\item{data}{Data to re-reference. Primarily meant for use with data of class
\code{eeg_data}.}

\item{...}{Further parameters to be passed to reref_eeg}

\item{ref_chans}{Channels to reference data to. Defaults to "average" i.e.
average of all electrodes in data. Character vector of channel names or numbers.}

\item{exclude}{Electrodes to exclude from average reference calculation.}

\item{robust}{Use median instead of mean; only used for average reference.}
}
\value{
object of class \code{eeg_data}, re-referenced as requested.
}
\description{
Used to reference the EEG data to a specified electrode or electrodes.
Defaults to average reference. When specific electrodes are used, they are
removed from the data. Meta-data about the referencing scheme is held in the
\code{eeg_data} structure.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method

\item \code{eeg_data}: Rereference objects of class \code{eeg_data}
}}

\examples{
# demo_epochs is average referenced by default
demo_epochs
# Rereference it but exclude B5 from calculation of the average
reref_eeg(demo_epochs, exclude = "B5")
# Reference data using the median of the reference channels rather than the mean
reref_eeg(demo_epochs, robust = TRUE)

}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
