% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencing.R
\name{reref_eeg}
\alias{reref_eeg}
\title{Referencing.}
\usage{
reref_eeg(data, ref_chans = "all")
}
\arguments{
\item{data}{Data to re-reference. Long format is expected.}

\item{ref_chans}{Channels to reference data to. Defaults to "all" i.e. average of all electrodes in data.}
}
\description{
Referencing.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
