% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{reref_eeg}
\alias{reref_eeg}
\title{Referencing}
\usage{
reref_eeg(data, ref_chans = "all")
}
\arguments{
\item{data}{Data to re-reference. Primarily meant for use with data of class \code{eeg_data}.}

\item{ref_chans}{Channels to reference data to. Defaults to "all" i.e. average of all electrodes in data.}
}
\description{
Used to reference the EEG data to specified electrode or electrodes. Defaults to average reference.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
