% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifiers.R
\name{rm_baseline}
\alias{rm_baseline}
\title{Baseline correction.}
\usage{
rm_baseline(data, time_lim = NULL)
}
\arguments{
\item{data}{Data to be baseline corrected.}

\item{time_lim}{Numeric character vector (e.g. time_lim <- c(-.1, 0)). If
none given, defaults to mean of whole epoch if the data is epoched, or the
channel mean if the data is continuous.}
}
\description{
Used to remove the mean of a specified time period from the data. Currently
only performs subtractive baseline. With a data frame, searches for
"electrode" and "epoch" columns, and groups on these when found. An electrode
column is always required; an epoch column is not.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
