% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_correction.R
\name{rm_baseline}
\alias{rm_baseline}
\alias{rm_baseline.eeg_data}
\alias{rm_baseline.eeg_epochs}
\alias{rm_baseline.data.frame}
\alias{rm_baseline.eeg_tfr}
\alias{rm_baseline.eeg_evoked}
\title{Baseline correction}
\usage{
rm_baseline(data, time_lim = NULL, ...)

\method{rm_baseline}{eeg_data}(data, time_lim = NULL, verbose = TRUE,
  ...)

\method{rm_baseline}{eeg_epochs}(data, time_lim = NULL, verbose = TRUE,
  ...)

\method{rm_baseline}{data.frame}(data, time_lim = NULL, verbose = TRUE,
  ...)

\method{rm_baseline}{eeg_tfr}(data, time_lim = NULL, type = "divide",
  verbose = TRUE, ...)

\method{rm_baseline}{eeg_evoked}(data, time_lim = NULL, verbose = TRUE,
  ...)
}
\arguments{
\item{data}{Data to be baseline corrected.}

\item{time_lim}{Numeric character vector (e.g. time_lim <- c(-.1, 0)). If
none given, defaults to mean of the whole of each epoch if the data is
epoched, or the channel mean if the data is continuous.}

\item{...}{other parameters to be passed to functions}

\item{verbose}{Defaults to TRUE. Output descriptive messages to console.}

\item{type}{Type of baseline correction to apply. Options are ("divide",
"ratio", "absolute", "db")}
}
\description{
Used to remove the mean of a specified time period from the data. Currently
only performs subtractive baseline. With a data frame, searches for
"electrode" and "epoch" columns, and groups on these when found. An electrode
column is always required; an epoch column is not.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: remove baseline from continuous \code{eeg_data}

\item \code{eeg_epochs}: Remove baseline from eeg_epochs

\item \code{data.frame}: Legacy method for data.frames

\item \code{eeg_tfr}: Method for \code{eeg_tfr} objects

\item \code{eeg_evoked}: Method for \code{eeg_evoked} objects
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
