% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{rotate_angle}
\alias{rotate_angle}
\title{Rotate channel locations}
\usage{
rotate_angle(chan_info, degrees)
}
\arguments{
\item{chan_info}{channel information structure}

\item{degrees}{degrees by which to rotate}
}
\description{
On import, channel locations may be rotated (e.g. Fpz pointing towards ears.)
}
\examples{

plot_electrodes(demo_epochs)
channels(demo_epochs) <- rotate_angle(channels(demo_epochs), 90)
plot_electrodes(demo_epochs)

}
