% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{run_ICA}
\alias{run_ICA}
\alias{run_ICA.eeg_epochs}
\title{Independent Component Analysis for EEG data}
\usage{
run_ICA(data, ...)

\method{run_ICA}{eeg_epochs}(data, method = "sobi", maxit = 1000,
  tol = 1e-06, pca = NULL, centre = TRUE, ...)
}
\arguments{
\item{data}{Data to be ICAed.}

\item{...}{Other parameters passed to function.}

\item{method}{"sobi" (default), "fastica", or "infomax".}

\item{maxit}{Maximum number of iterations of the Infomax and Fastica ICA
algorithms.}

\item{tol}{Convergence tolerance for fastica and infomax. Defaults to 1e-06.}

\item{pca}{Reduce the number of dimensions using PCA before running ICA.
Numeric,  >1 and < number of channels}

\item{centre}{Defaults to TRUE. Centre the data on zero by subtracting the column mean. See notes on usage.}
}
\value{
An \code{eeg_ICA} object containing an ICA decomposition
}
\description{
Performs Independent Component Analysis for EEG data. Currently only
available with on epoched data. Implements three different methods of ICA -
fastica, extended Infomax, and Second-Order Blind Identification (SOBI).
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Run ICA on an \code{eeg_epochs} object
}}

\section{Notes on ICA usage}{


  It is recommended to mean-centre your data appropriately before running
  ICA. The implementations of FASTICA and extended-Infomax from the `ica`
  package, and of SOBI ICA have this as an option which is enabled by
  default, while the implementation of FASTICA in the fICA package enforces
  mean-centring of the columns of the data. With epoched data, it is
  recommended to centre each epoch on zero, rather than centre on the overall
  channel mean. This can be achieved with the `rm_baseline()` function. SOBI
  ICA will do this automatically, whereas the other ICA implementations will
  centre on the channel means, not the epoch means.
}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
