% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{run_ICA}
\alias{run_ICA}
\alias{run_ICA.eeg_epochs}
\title{Independent Component Analysis for EEG data}
\usage{
run_ICA(data, ...)

\method{run_ICA}{eeg_epochs}(data, method = "sobi", maxit = 500,
  tol = 1e-06, ...)
}
\arguments{
\item{data}{Data to be ICAed.}

\item{...}{Other parameters passed to function.}

\item{method}{"sobi" (default), "fastica", or "infomax".}

\item{maxit}{Maximum number of iterations of the Infomax and Fastica ICA
algorithms.}

\item{tol}{Convergence tolerance for fastica and infomax. Defaults to 1e-06.}
}
\value{
An \code{eeg_ICA} object containing an ICA decomposition
}
\description{
Performs Independent Component Analysis for EEG data. Currently only
available with on epoched data. Implements three different methods of ICA -
fastica, extended Infomax, and Second-Order Blind Identification (SOBI).
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Run ICA on an \code{eeg_epochs} object
}}

\author{
Matt Craddock \email{matt@mattcraddock.com}
}
