% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{run_iir}
\alias{run_iir}
\title{Internal function for running IIR filtering}
\usage{
run_iir(data, low_freq = NULL, high_freq = NULL, filter_order, srate,
  plot_filt = FALSE, silent = FALSE)
}
\description{
Internal function for running IIR filtering
}
