% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{run_iir}
\alias{run_iir}
\title{Internal function for running IIR filtering}
\usage{
run_iir(
  data,
  low_freq = NULL,
  high_freq = NULL,
  filter_order,
  srate,
  silent = FALSE
)
}
\arguments{
\item{data}{Data to be filtered}

\item{low_freq}{Low passband edge.}

\item{high_freq}{High passband edge.}

\item{filter_order}{Order of the Butterworth filter.}

\item{srate}{Sampling rate of the signal.}

\item{silent}{Turns off filtering messages.}
}
\description{
Internal function for running IIR filtering
}
\keyword{internal}
