% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_elecs}
\alias{select_elecs}
\title{Select electrodes from a given dataset.}
\usage{
select_elecs(data, ...)
}
\arguments{
\item{data}{An EEG dataset.}

\item{electrode}{A character vector of electrode labels for selection or removal.}

\item{keep}{Defaults to TRUE. Set to false to *remove* the selected electrodes.}
}
\value{
Data frame with only data from the chosen electrodes
}
\description{
This is a generic function for selection of electrodes from an EEG dataset.
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
