% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_epochs}
\alias{select_epochs}
\alias{select_epochs.eeg_data}
\alias{select_epochs.eeg_epochs}
\title{Select epochs from eeg_data}
\usage{
select_epochs(data, ...)

\method{select_epochs}{eeg_data}(data, epoch_no = NULL, keep = TRUE,
  df_out = FALSE, ...)

\method{select_epochs}{eeg_epochs}(data, epoch_no = NULL,
  epoch_events = NULL, keep = TRUE, df_out = FALSE, ...)
}
\arguments{
\item{data}{\code{eeg_epochs} object from which to select epochs.}

\item{...}{Parameters passed to specific methods}

\item{epoch_events}{Select epochs containing any of the specified events.}
}
\description{
This is a generic function for selecting epochs from an epoched data set.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_data}: Select epochs from \code{eeg_data} objects.

\item \code{eeg_epochs}: Selection of epochs from eeg_epochs.
}}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
