% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_epochs}
\alias{select_epochs}
\alias{select_epochs.default}
\alias{select_epochs.eeg_epochs}
\alias{select_epochs.eeg_ICA}
\alias{select_epochs.eeg_tfr}
\title{Select epochs}
\usage{
select_epochs(data, ...)

\method{select_epochs}{default}(data, ...)

\method{select_epochs}{eeg_epochs}(
  data,
  epoch_events = NULL,
  epoch_no = NULL,
  keep = TRUE,
  df_out = FALSE,
  ...
)

\method{select_epochs}{eeg_ICA}(
  data,
  epoch_events = NULL,
  epoch_no = NULL,
  keep = TRUE,
  df_out = FALSE,
  ...
)

\method{select_epochs}{eeg_tfr}(
  data,
  epoch_events = NULL,
  epoch_no = NULL,
  keep = TRUE,
  df_out = FALSE,
  ...
)
}
\arguments{
\item{data}{\code{eeg_epochs} object from which to select epochs.}

\item{...}{Parameters passed to specific methods}

\item{epoch_events}{Select epochs containing any of the specified events. Can
be numeric or a character string. Will override any epoch_no input.}

\item{epoch_no}{Select epochs by epoch number.}

\item{keep}{Defaults to TRUE, meaning select the specified epochs. Set to
FALSE to remove specified epochs.}

\item{df_out}{Output a data.frame instead of an eeg_data object.}
}
\description{
This is a generic function for selecting epochs from an epoched data set.
}
\section{Methods (by class)}{
\itemize{
\item \code{select_epochs(default)}: Select from generic object

\item \code{select_epochs(eeg_epochs)}: Selection of epochs from \code{eeg_epochs} objects.

\item \code{select_epochs(eeg_ICA)}: Selection of epochs from \code{eeg_ICA} objects.

\item \code{select_epochs(eeg_tfr)}: Selection of epochs from \code{eeg_tfr} objects.

}}
\examples{
select_epochs(demo_epochs, epoch_no = 1:5)
demo_ica <- run_ICA(demo_epochs, pca = 10)
select_epochs(demo_ica, epoch_no = 1:5)

}
\seealso{
\code{\link[=select_times]{select_times()}} and \code{\link[=select_elecs]{select_elecs()}}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
\concept{data selection functions}
