% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_times}
\alias{select_times}
\title{Select timerange}
\usage{
select_times(data, ...)
}
\arguments{
\item{data}{Data from which to select}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Generic function for selecting specific time ranges from a given dataset.
Input can be a dataframe or an object of class \code{eeg_data}.
}
\seealso{
\code{\link{select_times.default}},
  \code{\link{select_times.eeg_data}}, \code{\link{select_elecs}}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
