% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection.R
\name{select_times}
\alias{select_times}
\alias{select_times.default}
\alias{select_times.eeg_data}
\alias{select_times.eeg_epochs}
\alias{select_times.eeg_evoked}
\alias{select_times.eeg_tfr}
\title{Select timerange}
\usage{
select_times(data, ...)

\method{select_times}{default}(data, time_lim = NULL, ...)

\method{select_times}{eeg_data}(data, time_lim = NULL, df_out = FALSE, ...)

\method{select_times}{eeg_epochs}(data, time_lim, df_out = FALSE, ...)

\method{select_times}{eeg_evoked}(data, time_lim, df_out = FALSE, ...)

\method{select_times}{eeg_tfr}(data, time_lim = NULL, df_out = FALSE, ...)
}
\arguments{
\item{data}{Data from which to select}

\item{...}{Further arguments passed to or from other methods.}

\item{time_lim}{A character vector of two numbers indicating the time range
to be selected e.g. c(min, max)}

\item{df_out}{Returns a data frame rather than an object of the same type
that was passed in.}
}
\value{
Data frame with only data from within the specified range.

\code{eeg_data} object
}
\description{
Generic function for selecting specific time ranges from a given dataset.
Input can be a dataframe, or an object of class \code{eeg_data},
\code{eeg_epochs}, or \code{eeg_evoked}. Note this finds the closest times to
those specified, so the time range returned may be slightly longer or shorter
than that requested.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default select times function

\item \code{eeg_data}: Select times from an eeg_data object

\item \code{eeg_epochs}: Select times in \code{eeg_epochs} objects

\item \code{eeg_evoked}: Select times in \code{eeg_evoked} objects

\item \code{eeg_tfr}: Select times from an \code{eeg_tfr} object
}}

\examples{
## Select timepoints from -.1 to .3
demo_epochs
short_epochs <- select_times(demo_epochs, time_lim = c(-.1, .3))
short_epochs

}
\seealso{
\code{\link{select_elecs}} and \code{\link{select_epochs}}

Other Data selection functions: 
\code{\link{select_elecs}()}
}
\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
\concept{Data selection functions}
