% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICA.R
\name{sobi_ICA}
\alias{sobi_ICA}
\title{SOBI ICA}
\usage{
sobi_ICA(data, maxiter)
}
\arguments{
\item{data}{Data to be ICAed.}

\item{maxiter}{Maximum number of iterations of the joint diagonalization}
}
\description{
Internal function for running SOBI ICA on an \code{eeg_epochs} object
}
\author{
A. Belouchrani and A. Cichocki. Adapted to R by Matt Craddock
  \email{matt@mattcraddock.com}
}
\keyword{internal}
