% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{sph_to_cart}
\alias{sph_to_cart}
\title{Convert spherical to cartesian 3d}
\usage{
sph_to_cart(theta, phi, r)
}
\arguments{
\item{theta}{should be in radians}

\item{phi}{should be in radians}

\item{r}{should be in radians}
}
\description{
Convert spherical to cartesian 3d
}
\keyword{internal}
