% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{sph_to_cart}
\alias{sph_to_cart}
\title{Convert spherical co-ordinates to Cartesian 3D co-ordinates}
\usage{
sph_to_cart(theta, phi, radius = 85)
}
\arguments{
\item{radius}{Radius of head (in mm)}

\item{sph_coords}{Theta and phi in degrees.}
}
\description{
Convert spherical co-ordinates to Cartesian 3D co-ordinates
}
\keyword{internal}
