% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_management.R
\name{sph_to_cart}
\alias{sph_to_cart}
\title{Convert spherical to cartesian 3d}
\usage{
sph_to_cart(theta, phi, radius)
}
\arguments{
\item{theta}{should be in degrees}

\item{phi}{should be in degrees}

\item{r}{should be in degrees}
}
\description{
Note that phi follows EEGLAB conventions (i.e. needs correcting to 90 - phi in some cases)
}
\keyword{internal}
