% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_source_density.R
\name{spheric_spline}
\alias{spheric_spline}
\title{Calculate a spherical spline smooth for interpolation of electrodes}
\usage{
spheric_spline(good_elecs, all_elecs)
}
\arguments{
\item{good_elecs}{Electrode positions that do not need interpolation}

\item{all_elecs}{Electrode positions including those to be interpolated}
}
\description{
Calculate a spherical spline smooth for interpolation of electrodes
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\keyword{internal}
