% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{split_vec}
\alias{split_vec}
\title{Split data into segments for Welch PSD .}
\usage{
split_vec(vec, seg_length, overlap)
}
\arguments{
\item{vec}{Data vector to be split up}
}
\description{
Split data into segments for Welch PSD .
}
\author{
Matt Craddock \mail{matt@mattcraddock.com}
}
