% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-extensions.R
\name{stat_scalpmap}
\alias{stat_scalpmap}
\alias{geom_head}
\alias{geom_mask}
\alias{geom_ears}
\alias{geom_channels}
\title{Create an interpolated scalp surface}
\usage{
stat_scalpmap(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  grid_res = 300, interpolate = FALSE, interp_limit = "skirt", ...)

geom_head(mapping = NULL, data = NULL, show.legend = NA,
  na.rm = TRUE, inherit.aes = TRUE, interp_limit = "skirt",
  r = NULL, ...)

geom_mask(mapping = NULL, data = NULL, show.legend = NA,
  na.rm = FALSE, colour = "white", size = rel(6.5),
  scale_fac = 1.32, ...)

geom_ears(mapping = NULL, data = NULL, show.legend = NA,
  na.rm = FALSE, ...)

geom_channels(mapping = NULL, data = NULL, geom = "point",
  show.legend = NA, inherit.aes = TRUE, na.rm = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{grid_res}{Resolution of the interpolation grid. (Defaults to 100
points).}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{interp_limit}{Topoplot with a "skirt" or inside the "head".}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{r}{Radius of head}

\item{colour}{For \code{geom_mask}, colour of the masking ring.}

\item{size}{For \code{geom_mask}, width of the masking ring.}

\item{scale_fac}{The radius of the ring is determined from the front-most
electrode's location by a scaling factor. Defaults to 1.32 * max(y).}

\item{geom}{"point" for points or "text" for labels. Default is "point".}
}
\description{
\code{stat_scalpmap} creates an interpolated surface for an irregular set of
x-y coordinates, as is typically required for a topographical EEG plot. Since
the surface should be approximately round, the function attempts to blank out
portions of the surface that lay outside the area within the electrodes.

\code{geom_head()} adds a headshape to a plot.

\code{geom_mask()} adds a masking ring to smooth the edges of a scalp map
generated by \code{stat_scalpmap()}, to give it a circular appearance.

\code{geom_ears} simply draws a pair of ears attached to the head shape.

\code{geom_channels} adds either points or text labels at channel locations.
This is a convenience function to prevent overplotting when the input data
contains many rows of data.
}
\seealso{
Other topoplot functions: \code{\link{geom_topo}}

Other topoplot functions: \code{\link{geom_topo}}

Other topoplot functions: \code{\link{geom_topo}}
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
\concept{topoplot functions}
