% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_format}
\alias{switch_format}
\title{Switch from wide to long format.}
\usage{
switch_format(x)
}
\arguments{
\item{data}{Data to convert}
}
\description{
Switch from wide to long format.
}
\author{
Matt Craddock \email{m.p.craddock@leeds.ac.uk}
}
