% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfr_analysis.R
\name{tf_morlet}
\alias{tf_morlet}
\title{Perform Morlet time-frequency analysis}
\usage{
tf_morlet(
  data,
  foi,
  n_freq,
  n_cycles,
  keep_trials,
  output,
  downsample,
  demean = TRUE,
  lang = "R",
  verbose
)
}
\arguments{
\item{data}{Data in \code{eeg_epochs} format.}

\item{foi}{Frequencies of interest. Scalar or character vector of the lowest
and highest frequency to resolve.}

\item{n_freq}{Number of frequencies to be resolved.}

\item{n_cycles}{Number of cycles at each frequency.}

\item{keep_trials}{Keep single trials or average over them before returning.}

\item{output}{Sets whether output is power, phase, or fourier coefficients.}

\item{downsample}{Downsampling factor (integer).}

\item{demean}{Remove mean before transforming.}

\item{lang}{defaults to R}

\item{verbose}{Print informative messages in console.}
}
\description{
Internal function for performing Morlet wavelet transforms using convolution
in frequency domain
}
\keyword{internal}
