% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{tf_morlet}
\alias{tf_morlet}
\alias{tf_morlet.eeg_epochs}
\title{Time-frequency analysis}
\usage{
tf_morlet(data, ...)

\method{tf_morlet}{eeg_epochs}(data, foi, n_freq, n_cycles, ...)
}
\arguments{
\item{data}{EEG data to be TF transformed}

\item{...}{Further parameters of the timefreq transformation}

\item{foi}{Frequencies of interest. Scalar or character vector of the lowest and highest frequency to resolve.}

\item{n_freq}{Number of frequencies to be resolved.}

\item{n_cycles}{Number of cycles at each frequency.}
}
\description{
Morlet wavelet time-frequency analysis.
}
\section{Methods (by class)}{
\itemize{
\item \code{eeg_epochs}: Time-frequency decomposition of \code{eeg_epochs} object.
}}

