% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot}
\alias{topoplot}
\title{Topographical Plotting Function for EEG}
\usage{
topoplot(df, electrodeLocs = NULL, method = "Biharmonic", rmax = 0.75,
  gridRes = 67)
}
\arguments{
\item{df}{An EEG dataset. Must have columns x, y, and amplitude at present. x and y are (Cartesian) electrode co-ordinates), amplitude is amplitude.}

\item{electrodeLocs}{A set of electrode locations.Can be in polar or spherical format. Will attempt to use 10-20 if not supplied.}

\item{method}{Interpolation method. Defaults to Biharmonic spline interpolation, which is the only one currently available.}

\item{rmax}{Maximum head radius if x is too small.}

\item{gridRes}{Resolution of the interpolated grid. Higher = smoother but slower.}
}
\description{
Allows simple plotting of functional data
}
