% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoplot.R
\name{topoplot}
\alias{topoplot}
\alias{topoplot.data.frame}
\alias{topoplot.eeg_data}
\alias{topoplot.eeg_ICA}
\title{Topographical Plotting Function for EEG}
\usage{
topoplot(data, ...)

\method{topoplot}{data.frame}(data, time_lim = NULL, limits = NULL,
  chanLocs = NULL, method = "Biharmonic", r = NULL, grid_res = 67,
  palette = "RdBu", interp_limit = "skirt", contour = TRUE,
  chan_marker = "point", quantity = "amplitude", montage = NULL,
  colourmap, ...)

\method{topoplot}{eeg_data}(data, time_lim = NULL, limits = NULL,
  chanLocs = NULL, method = "Biharmonic", r = NULL, grid_res = 67,
  palette = "RdBu", interp_limit = "skirt", contour = TRUE,
  chan_marker = "point", quantity = "amplitude", montage = NULL, ...)

\method{topoplot}{eeg_ICA}(data, time_lim = NULL, limits = NULL,
  chanLocs = NULL, method = "Biharmonic", r = NULL, grid_res = 67,
  palette = "RdBu", interp_limit = "skirt", contour = TRUE,
  chan_marker = "point", quantity = "amplitude", montage = NULL,
  colourmap, comp, ...)
}
\arguments{
\item{data}{An EEG dataset. Must have columns x, y, and amplitude at present.
x and y are (Cartesian) electrode co-ordinates), amplitude is amplitude.}

\item{...}{Various arguments passed to specific functions}

\item{time_lim}{Timepoint(s) to plot. Can be one time or a range to average
over. If none is supplied, the function will average across all timepoints
in the supplied data.}

\item{limits}{Limits of the fill scale - should be given as a character vector
with two values specifying the start and endpoints e.g. limits = c(-2,-2).
Will ignore anything else. Defaults to the range of the data.}

\item{chanLocs}{Allows passing of channel locations (see \code{electrode_locations})}

\item{method}{Interpolation method. "Biharmonic" or "gam". "Biharmonic"
implements the same method used in Matlab's EEGLAB. "gam" fits a
Generalized Additive Model with k = 40 knots. Defaults to biharmonic spline
interpolation.}

\item{r}{Radius of cartoon head_shape; if not given, defaults to 1.1 * the
maximum y electrode location.}

\item{grid_res}{Resolution of the interpolated grid. Higher = smoother but
slower.}

\item{palette}{Defaults to RdBu if none supplied. Can be any from
RColorBrewer or viridis. If an unsupported palette is specified, switches
to Greens.}

\item{interp_limit}{"skirt" or "head". Defaults to "skirt". "skirt"
interpolates just past the farthest electrode and does not respect the
boundary of the head_shape. "head" interpolates up to the radius of the
plotted head.}

\item{contour}{Plot contour lines on topography (defaults to TRUE)}

\item{chan_marker}{Set marker for electrode locations. "point" = point,
"name" = electrode name, "none" = no marker. Defaults to "point".}

\item{quantity}{Allows plotting of arbitrary quantitative column. Defaults to
amplitude. Use quoted column names. E.g. "p.value", "t_statistic".}

\item{montage}{Name of an existing montage set. Defaults to NULL; (currently
only 'biosemi64alpha' available other than default 10/20 system)}

\item{colourmap}{Deprecated, use palette instead.}

\item{comp}{Component to plot (numeric)}
}
\description{
Allows topographical plotting of functional data. Output is a ggplot2 object.

The functions works for both standard data frames and objects of class
\code{eeg_data}.

Both \code{eeg_epochs} and \code{eeg_data} objects are supported.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Topographical plotting of data.frames and other non
eeg_data objects.

\item \code{eeg_data}: Topographical plotting of \code{eeg_data} objects.

\item \code{eeg_ICA}: Topographical plot for \code{eeg_ICA} objects
}}

\section{Notes on usage of Generalized Additive Models for interpolation}{
 The
  function fits a GAM using the gam function from mgcv. Specifically, it fits
  a spline using the model function gam(z ~ s(x, y, bs = "ts", k = 40). Using
  GAMs for smooths is very much experimental. The surface is produced from
  the predictions of the GAM model fitted to the supplied data. Values at
  each electrode do not necessarily match actual values in the data:
  high-frequency variation will tend to be smoothed out. Thus, the method
  should be used with caution.
}

\author{
Matt Craddock, \email{matt@mattcraddock.com}
}
