% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefact_viewer.R
\name{view_artefacts}
\alias{view_artefacts}
\title{Artefact browser}
\usage{
view_artefacts(data)
}
\description{
An interactive Shiny app that allows exploration of channel and epoch
statistics.
}
\keyword{internal}
